% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_helpers.R
\name{categorical_mode}
\alias{categorical_mode}
\title{Find the mode of a categorical variable}
\usage{
categorical_mode(x)
}
\arguments{
\item{x}{vector to find mode of}
}
\value{
the most commonly occouring character
}
\description{
Find the mode of a categorical variable
}
\examples{
#categorical_mode(c("Cat", "Cat", "Monkey"))

}
\keyword{internal}
