% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cellzscope.R
\name{cellzscope_import}
\alias{cellzscope_import}
\title{Import cellSZcope data}
\usage{
cellzscope_import(raw, model, key, experimentname)
}
\arguments{
\item{raw}{File location of the raw dataset}

\item{model}{File locaiton of the modeled dataset}

\item{key}{Location of a vascar standard lookup table. Optional, but can be used to import more granular data than possible with the built in row names}
}
\value{
a standard vascar dataset
}
\description{
Data must first be exported in two parts from the cellZScope software. This can then be imported to a standard vascar dataset with this function
}
\examples{

model = system.file("extdata/mdckmodel.txt", package = "vascr")
raw = system.file("extdata/mdckspectra.txt", package = "vascr")
key = system.file("extdata/mdckkey.csv", package = "vascr")

alldata = cellzscope_import(raw, model)
vascr_plot(alldata, unit = "TER", preprocessed = TRUE, frequency = 0,
 time = c(0,50), replication = "experiments")

alldatakey = cellzscope_import(raw, model, key)
alldatakey$Instrument = "CZSII"
vascr_plot(alldatakey, unit = "TER", frequency = 0, time = c(0,50), 
replication = "experiments", preprocessed = TRUE)

}
