% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cellzscope.R
\name{cellzscope_import_model}
\alias{cellzscope_import_model}
\title{Import modeled cellZScope data}
\usage{
cellzscope_import_model(model, key)
}
\arguments{
\item{model}{The location of modeled data exported from the cellZscope data}

\item{key}{Optional, a keyfile containing data on which well contains which sample}
}
\value{
A vascar compatable dataset.
}
\description{
Imports modeled data previously exported from the cellZscope package. Will not assign names unless a keyfile or raw dataset is provided, as no well naming data is provided in this filetype.
}
