% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_continuous.R
\name{col_exists}
\alias{col_exists}
\title{Check if a column exists in a dataset}
\usage{
col_exists(data, col)
}
\arguments{
\item{data}{The dataset to check}

\item{col}{The column to check. Should be a character string}
}
\value{
A boolean, true if the column is found, false if not
}
\description{
Check if a column exists in a dataset
}
\examples{
#col_exists(growth.df, "Vector")
#col_exists(growth.df, "Unit")
}
\keyword{internal}
