% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{comma_to_numeric}
\alias{comma_to_numeric}
\title{Convert comma containing values to numeric ones}
\usage{
comma_to_numeric(comma_array)
}
\arguments{
\item{comma_array}{The array (and hence column of a data frame also works) containing comma separated values to process}
}
\value{
An array containing numeric data. NA will be returned if not numeric, and a warning will be generated
}
\description{
This function converts numbers formatted with commas, such as those used in text variables to numbers. This process is lossy, so is used as late as possible before a continuous variable is needed. Preserves decimal places. Vectorised.
}
\examples{


}
