% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data.df}
\alias{data.df}
\title{Example data set from an ECIS growth experiment}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Experiment}{Date the experiment was conducted}
  \item{Frequency}{Frequency of colleciton, modeled units have a frequency of 0}
  \item{Sample}{Name of the sample located in that well}
  \item{Time}{Capture time}
  \item{TimeID}{Sequential number of the time point captured}
  \item{Unit}{Unit that the value is measured in}
  \item{Value}{Numerical value in units}
  \item{Well}{Well of plate that was measured}
  ...
}}
\source{
James Hucklesby
}
\usage{
data.df
}
\description{
A dataset containing growth curves from a variety of experiments
}
\keyword{datasets}
