% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_96_well_names}
\alias{ecis_96_well_names}
\title{All the well names of a 96 well plate}
\usage{
ecis_96_well_names()
}
\value{
Vector containing all wells of a 96 well plate
}
\description{
All the well names of a 96 well plate
}
\examples{
ecis_96_well_names()

}
