% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_align_key}
\alias{ecis_align_key}
\title{Align key points in an ECIS trace}
\usage{
ecis_align_key(data.df, point, discrepancy = 5)
}
\arguments{
\item{data.df}{A standard ECIS data file}

\item{point}{Which key point, either 'max' or 'min'}

\item{discrepancy}{A standard rounding constant to compensate for rounding errors in the subtraciton process}
}
\value{
An ECIS dataset where the key time points all happen at time 0
}
\description{
This will either align the max or minimum points from each graph as specified
}
\details{
Sets the time at which each replicate well is maximal to time 0. Results in variables aligned by maximum time, rather than time from seeding.
}
\examples{

data = ecis_align_key(growth.df, 'max')
head(data)
data = ecis_align_key(growth.df, 'min')
head(data)
}
