% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{ecis_align_max}
\alias{ecis_align_max}
\title{Align maximum points in an ECIS trace}
\usage{
ecis_align_max(data.df, includemaxima = FALSE)
}
\arguments{
\item{data.df}{}

\item{includemaxima}{}
}
\value{

}
\description{
Align maximum points in an ECIS trace
}
