% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_animatefrequency}
\alias{ecis_animatefrequency}
\title{Title}
\usage{
ecis_animatefrequency(alldata.df, unittoplot, frames)
}
\arguments{
\item{alldata.df}{}

\item{unittoplot}{}

\item{frames}{}
}
\description{
Title
}
