% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_methods.R
\name{ecis_ANOVA}
\alias{ecis_ANOVA}
\title{ANOVA analyasis of a single ECIS timepoint}
\usage{
ecis_ANOVA(data.df, unit, frequency, time, posthoc = "bonferoni")
}
\arguments{
\item{data.df}{A standard ECIS dataset.}

\item{unit}{Unit to be analysed.}

\item{frequency}{Frequency to be used. All modeled units have a frequency of 0.}

\item{time}{Time to be used.}

\item{posthoc}{Post hoc test to use. Options include Tukey. More to be added in the future.}
}
\value{
ANOVA results printed to the console.
}
\description{
This function takes a standard ECIS dataset and runs two way ANOVA on it. It then runs a post-hoc test to determine which groups are significatnly difference. Results are printed to the console as text.
}
\examples{
ecis_ANOVA(ECISR::data.df, "Rb",0,75)

}
