% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_experiments.R
\name{ecis_check_integrity}
\alias{ecis_check_integrity}
\title{Title}
\usage{
ecis_check_integrity(data.df, verbose = FALSE)
}
\arguments{
\item{data.df}{An ECIS dataframe, subset if needed}
}
\value{
A whole slew of tests
}
\description{
Title
}
\examples{
# Prep a known good compliment of data
data = ecis_subset(growth.df, time = 100, unit = "R", frequency = 4000)

# Test a full compliment of data passes
ecis_check_integrity(data)

ecis_check_integrity(growth.df)
# Test function picks up unbalanced replicate #'s
data2 = ecis_exclude(data, well = c("A1", "B1", "C1"))
ecis_check_integrity(data2)
# Test function picks up missing pairs (in this case due to Rb not being established)
data3 = ecis_subset(growth.df, time = 100, unit = "Rb")
ecis_check_integrity(data3)


}
