% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{ecis_cols}
\alias{ecis_cols}
\title{Return the ECIS cols used in this package}
\usage{
ecis_cols(set = "core")
}
\arguments{
\item{set}{The set of columns to request. Default is core.}
}
\value{
A vector of the columns requested
}
\description{
Return the ECIS cols used in this package
}
\examples{

ecis_cols()

}
