% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmaster.R
\name{ecis_combine}
\alias{ecis_combine}
\title{Title}
\usage{
ecis_combine(...)
}
\arguments{
\item{...}{List of data frames to be combined}
}
\value{
A single data frame containing all the data imported, automaticaly incremented by experiment
}
\description{
Title
}
\examples{


}
