% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_currentfrequency}
\alias{ecis_currentfrequency}
\title{Current aquisition rate}
\usage{
ecis_currentfrequency(data.df)
}
\arguments{
\item{data.df}{The dataframe to compute the current data aquisition frequency of}
}
\value{
The current aquisition rate of the data frame
}
\description{
Current aquisition rate
}
\examples{

ecis_currentfrequency(data.df)


}
