% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_default}
\alias{ecis_default}
\title{Look up the default variable values for all supported instruments}
\usage{
ecis_default(data)
}
\arguments{
\item{data}{The dataset to derrive the instrument type from}
}
\value{

}
\description{
Central source of lookup values for default graph settings for various instruments.
}
\examples{
ecis_default(growth.df)

}
