% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_detect_badwells}
\alias{ecis_detect_badwells}
\title{Automatically strip badly connected wells from an ECIS dataset}
\usage{
ecis_detect_badwells(data.df, threshold = 5, frequency = 4000, unit = "R")
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{threshold}{How stringent to be in excluding wells. Higher is less stringent. Default is 5.}

\item{frequency}{Frequency to run numbers on, default is 4000}

\item{unit}{Unit to use in detection, default is R}
}
\value{
A tibble containing the offending wells, which experiment they are from and the score they were removed with
}
\description{
Automatically strip badly connected wells from an ECIS dataset
}
\examples{
# Make a defective well in the dataset
baddata = growth.df
welltobreak = "B1"
baddata$randoms = sample(baddata$Value*2, size = nrow(baddata), replace = TRUE)
baddata$Value = baddata$Value + ((baddata$Well == welltobreak)*baddata$randoms)

welltobreak = "H4"
baddata$randoms = sample(baddata$Value*2, size = nrow(baddata), replace = TRUE)
baddata$Value = baddata$Value + ((baddata$Well == welltobreak)*baddata$randoms)

# Plot out the well, and then try to detect it
ecis_detect_badwells(baddata,1)

# Check it works for a good dataset
ecis_detect_badwells(growth.df,1)

}
