% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_detect_normal}
\alias{ecis_detect_normal}
\title{Detect if an ECIS dataset has been normalised}
\usage{
ecis_detect_normal(data.df)
}
\arguments{
\item{data.df}{an ECIS dataset}
}
\value{
The time the data was normalised to, or FALSE if not normalised
}
\description{
Detect if an ECIS dataset has been normalised
}
\examples{

standard = growth.df
normal = ecis_normalise(growth.df, 100)
ecis_detect_normal(standard)
ecis_detect_normal(normal)
}
