% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_continuous.R
\name{ecis_detect_vehicle}
\alias{ecis_detect_vehicle}
\title{Detect vehicles in an ecisr dataset}
\usage{
ecis_detect_vehicle(data, force_replace = FALSE)
}
\arguments{
\item{data}{An ecisr dataset}

\item{force_replace}{Should vehicles be replaced, if they already exist. True replaces the column, false does not. Default FALSE.}
}
\value{
An ecisr dataset with an extra IsVehicleControl column
}
\description{
Detect vehicles in an ecisr dataset
}
\examples{
# data = xcell
# ecis_detect_vehicle(data)

}
