% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_continuous.R
\name{ecis_detect_vehicle_control_wells}
\alias{ecis_detect_vehicle_control_wells}
\title{Detect the wells R thinks contains vehicles}
\usage{
ecis_detect_vehicle_control_wells(data, force_replace = FALSE)
}
\arguments{
\item{data}{The datset to search}

\item{force_replace}{Should existing IsVehicleControl wells be replaced (if present)?}
}
\value{

}
\description{
Convenience function. This first runs ecis_detect_vehicle on the dataset, then provides a summary table of which wells are associated with which vehicle. Provides a usefull check that vehicle detection has been preformed correctly.
}
\examples{
#ecis_detect_vehicle_control_wells(data)

}
