% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_exclude}
\alias{ecis_exclude}
\title{Exclude erronious data from an ECIS dataframe}
\usage{
ecis_exclude(data.df, samples = FALSE, wells = FALSE,
  experiments = FALSE, times = FALSE, values = FALSE, vars = FALSE,
  vals = FALSE, vs = FALSE)
}
\arguments{
\item{data.df}{The source dataset}

\item{samples}{The sample(s) to exclude}

\item{wells}{The well(s) to exclude}

\item{experiments}{The experiment(s) to exclude}

\item{times}{The time(s) to exclude}

\item{values}{The value(s) to exclude}

\item{vars}{The isolated variables to exclude}

\item{vals}{The isolated values to exclude}

\item{vs}{The isolated variables-unit pairs to exclude}
}
\value{
The altered dataset
}
\description{
Exclude erronious data from an ECIS dataframe
}
\examples{

unique(growth.df$Sample)
excludedgrowth.df = ecis_exclude(growth.df, samples = c("35,000 cells", "0 cells"))
unique(excludedgrowth.df$Sample)

unique(growth.df$Well)
excludedgrowth.df = ecis_exclude(growth.df, wells = c("A1", "B1", "C1"))
unique(excludedgrowth.df$Well)

unique(growth.df$Experiment)
excludedgrowth.df = ecis_exclude(growth.df, experiment = c(1,2))
unique(excludedgrowth.df$Experiment)

}
