% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_exclude_badwells}
\alias{ecis_exclude_badwells}
\title{Exclude automatically detected wells that have a connection issue from the dataset}
\usage{
ecis_exclude_badwells(data.df, threshold = 5, frequency = 4000,
  unit = "R", verbose = TRUE)
}
\arguments{
\item{data.df}{The dataset to parse}

\item{threshold}{The threshold stringency to use in detection. Default is 5, the range of 1-10 may be appropriate. Higher numbers are less stringent.}

\item{frequency}{The frequency to use for detection, default is 4000 Hz}

\item{unit}{The unit to run the detection on, default is R}

\item{verbose}{Prints which wells have been removed in the terminal. Should be used when first investigating data to allow for follow up plots with ecis_isolate_well to be conducted.}
}
\value{
A standard ECIS dataframe, minus the detected wells
}
\description{
Exclude automatically detected wells that have a connection issue from the dataset
}
\examples{
# Make a defective well in the dataset
baddata = growth.df
welltobreak = "B1"
baddata$randoms = sample(baddata$Value*2, size = nrow(data), replace = TRUE)
baddata$Value = baddata$Value + ((baddata$Well == welltobreak)*baddata$randoms)

welltobreak = "H4"
baddata$randoms = sample(baddata$Value*2, size = nrow(data), replace = TRUE)
baddata$Value = baddata$Value + ((baddata$Well == welltobreak)*baddata$randoms)

ecis_exclude_badwells(baddata, threshold = 1)
ecis_exclude_badwells(growth.df, threshold = 1)

}
