% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_explode_continuous}
\alias{ecis_explode_continuous}
\title{Title}
\usage{
ecis_explode_continuous(data.df, fields)
}
\arguments{
\item{data.df}{}

\item{fields}{}
}
\description{
Title
}
\examples{
data.df = growth.df
data.df$Sample = paste(data.df$Sample, " 10nm nothing")
ecis_explode_continuous(data.df, c("Cells", "Nothing"))

}
