% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{ecis_exploded_cols}
\alias{ecis_exploded_cols}
\title{Return the exploded cols in a dataset}
\usage{
ecis_exploded_cols(data)
}
\arguments{
\item{data}{The ECIS dataset with exploded columns}
}
\value{
A vector of the returned columns
}
\description{
Return the exploded cols in a dataset
}
\examples{

growth.df$Instrument = "ECIS"
ecis_explode(growth.df)

}
