% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_factorise_and_sort}
\alias{ecis_factorise_and_sort}
\title{Factorise and sort a the column of a data frame for plotting}
\usage{
ecis_factorise_and_sort(data, sortkeyincreasing = TRUE)
}
\arguments{
\item{data}{}

\item{sortkeyincreasing}{}
}
\value{

}
\description{
Factorise and sort a the column of a data frame for plotting
}
