% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_find_frequency}
\alias{ecis_find_frequency}
\title{Align frequencies}
\usage{
ecis_find_frequency(data.df, frequency)
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{frequency}{The tfrequency that needs rounding}
}
\value{
A timepoint that exactly aligns with a measured datapoint
}
\description{
When running analyasis, you can only subset or plot a time that exists in the dataset. These are not always logical or easy to remember. This function rounds the number given to the nearest frequency that is actually in the dataset.
}
\examples{
ecis_find_frequency(growth.df, 4382)

}
