% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_fix_error}
\alias{ecis_fix_error}
\title{Fix typographical errors in an ECIS dataframe}
\usage{
ecis_fix_error(data.df, incorrect, correct, limit = "None")
}
\arguments{
\item{data.df}{the data to correct}

\item{incorrect}{the error to be corrected}

\item{correct}{the corrrect string}

\item{limit}{can be set to "Sample or "Experiment" to limit the replacement to a single variable}
}
\value{
the repared dataframe
}
\description{
This function will go through all the sample and experiment names and replace all the values that are incorrect. Usefull for fixing up typographical errors, or places where you have named things differently inadvertently.
}
\examples{

ecis_fix_error(growth.df, "cells", "cells plated")

}
