% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_generate_continuous}
\alias{ecis_generate_continuous}
\title{Title}
\usage{
ecis_generate_continuous(string)
}
\arguments{
\item{string}{}
}
\description{
Title
}
\examples{

ecis_generate_continuous("  30,000.01 cells + 1 mg/ml TNFa")

}
