% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{ecis_implode_continuous}
\alias{ecis_implode_continuous}
\title{Implode a continuous dataset}
\usage{
ecis_implode_continuous(data, cols = NULL, stripidentical = FALSE)
}
\arguments{
\item{data}{A standard ECIS data frame}

\item{cols}{The columns to plot. Default of null plots the whole lot.}

\item{stripidentical}{Strip out columns that are identical in all samples. Tidies up graphs significantly (at the cost of being less explicit). Default is FALSE as the user must manually specify what has been removed in the graph title, or data is implicity lost.}
}
\value{
the cleaned up data frame
}
\description{
This function implodes a coninuous dataset to re-create Sample names from only some (or all) of the exploded columns. Provides a better way to subset data for plotting (as it allows for exclusion of values that are always the same) and for the generation of continuous plots.
}
\examples{

# Make a test dataset
library(stringr)
data.df = growth.df
data.df$Sample = str_replace(data.df$Sample, " ", "_")
data.df$Sample = paste(data.df$Sample, "+ 10_nm nothing important")
data.df$Sample = paste(data.df$Sample, "+ 4_nm Carpet + ECV_line")
#exploded = ecis_explode(data.df)

#Check the function
#imploded = ecis_implode_continuous(exploded, cols = c('cells', 'nm nothing important'))
#imploded = ecis_implode_continuous(exploded, cols = c('cells', "nm nothing important"), stripidentical = TRUE)

}
