% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmaster.R
\name{ecis_import_long}
\alias{ecis_import_long}
\title{Import all ECIS values}
\usage{
ecis_import_long(resample, modeled, key)
}
\arguments{
\item{resample}{A resampled APB file for import}

\item{modeled}{A modeled APB file for import}

\item{key}{A CSV file containing each well and it's corresponding sample}
}
\description{
Import all ECIS values
}
\examples{


}
