% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ecis_import_model}
\alias{ecis_import_model}
\title{Import raw modeled data}
\usage{
ecis_import_model(modeled.csv, sampledefine)
}
\arguments{
\item{modeled.csv}{Raw modeled data in APB format}

\item{sampledefine}{CSV file containing which wells correspond to which values}
}
\value{
Data frame containing modeled data
}
\description{
Import raw modeled data
}
\examples{

 #First determine the locatins of your files relative to your dataset.
 #Here we use system.file to pull a default out of the filesystem, but you 
 #can use a path relative to the file you are working on. 
 #E.G 'Experiment1/Raw.abp'

rawdata = system.file('Model.csv', package = 'ecisr')
sampledefine = system.file('Samples.csv', package = 'ecisr')

#Then run the import

data = ecis_import_model(rawdata, sampledefine)
head(data)

}
