% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmaster.R
\name{ecis_import_model_long}
\alias{ecis_import_model_long}
\title{Title}
\usage{
ecis_import_model_long(rawdata, samples)
}
\arguments{
\item{rawdata}{Raw modeled data in APB format}

\item{samples}{CSV file containing which wells correspond to which values}
}
\value{
Data frame containing modeled data
}
\description{
Title
}
\examples{

}
