% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ecis_import_raw}
\alias{ecis_import_raw}
\title{ECIS raw data importer}
\usage{
ecis_import_raw(rawdata, sampledefine, by, from = Inf, to = Inf,
  zero_time = 0, verbose = TRUE, no_process = FALSE)
}
\arguments{
\item{rawdata}{A resampled ABP file containing the un-modeled data}

\item{sampledefine}{A CSV file containing well numbers and their corresponding sample names}
}
\value{
Data frame containing all the raw data readings from the ECIS Z0 instrument
}
\description{
Raw data importer, generates a r dataframe from a raw ABP file
}
\examples{

#First determine the locatins of your files relative to your dataset. 
#Here we use system.file to pull a default out of the filesystem, 
#but you can use a path relative to the file you are working on. 
#E.G 'Experiment1/Raw.abp'

rawdata = system.file('Resample.abp', package = 'ecisr')
sampledefine = system.file('Samples.csv', package = 'ecisr')

#Then run the import

data2 = ecis_import_raw(rawdata, sampledefine, .1, 1, 1.4)
head(data)

data2 = ecis_import_raw(rawdata, NULL, 1, 0, 70)
data3 = subset(data2, Value<20000)
ecis_plot(data3, replication = "plate")

}
