% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmaster.R
\name{ecis_import_raw_long}
\alias{ecis_import_raw_long}
\title{Raw data importer}
\usage{
ecis_import_raw_long(rawdata, sampledefine)
}
\arguments{
\item{rawdata}{A resampled ABP file containing the un-modeled data}

\item{sampledefine}{A CSV file containing well numbers and their corresponding sample names}
}
\value{
Data frame containing all the raw data readings from the ECIS Z0 instrument
}
\description{
Raw data importer
}
\examples{

#ecis_import_raw_long(resampled.abp, sampledefinitions.csv)

}
