% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_is_modeled_unit}
\alias{ecis_is_modeled_unit}
\title{Check if a selected unit is modelled}
\usage{
ecis_is_modeled_unit(unit)
}
\arguments{
\item{unit}{The ecisr symbol for the unit}
}
\value{
A boolean, true if it is modelled, false if it is raw electrical data
}
\description{
Check if a selected unit is modelled
}
