% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_isolate_well}
\alias{ecis_isolate_well}
\title{Graph a single well that is misbehaving}
\usage{
ecis_isolate_well(data.df, well)
}
\arguments{
\item{data.df}{}

\item{well}{}
}
\value{
A ggplot graph showing the isoated well and the experimental median well
}
\description{
Graph a single well that is misbehaving
}
\examples{

ecis_isolate_well(growth.df, "A3")

}
