% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_methods.R
\name{ecis_make_labeltable}
\alias{ecis_make_labeltable}
\title{Make a dataframe of what is significant}
\usage{
ecis_make_labeltable(data.df, time, unit, frequency, confidence = 0.95,
  format = "toplot")
}
\arguments{
\item{data.df}{The dataframe to analyse}

\item{time}{The time to analyse}

\item{unit}{The unit to analyse}

\item{frequency}{The frequency to analyse}

\item{confidence}{The confidence level to analyse - default is 0.95}

\item{format}{The format to return the data frame as}
}
\value{
A table of what is significant
}
\description{
Make a dataframe of what is significant
}
\examples{

ecis_make_labeltable(growth.df, 50, "R", 4000, 0.95)


}
