% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{ecis_normalise}
\alias{ecis_normalise}
\title{Normalise ECIS data to a single time point}
\usage{
ecis_normalise(data.df, normtime, divide = FALSE)
}
\arguments{
\item{normtime}{Time to normalise the data to}

\item{divide}{If set to true, data will be normalsed via a division. If set to false (default) data will be normalsed by subtraction.}

\item{data}{Standard ECIS Dataframe}
}
\value{
A standard ECIS dataset with each value normalised
}
\description{
This function normalises each unique experiment/well combination to it's value at the specified time. Contains options to do this either by division or normalisation. Can be run twice if both operations are desired.
}
\examples{

ecis_normalise(data.df, 100)

}
