% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot}
\alias{ecis_plot}
\title{ECIS plot}
\usage{
ecis_plot(data, unit, frequency, replication, time = Inf, samplesubset)
}
\arguments{
\item{data}{}

\item{unit}{}

\item{frequency}{}

\item{replication}{}

\item{time}{}
}
\description{
ECIS plot
}
