% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot_all}
\alias{ecis_plot_all}
\title{Plot all replicates from multiple experiments individualy}
\usage{
ecis_plot_all(data.df, unit, frequency)
}
\arguments{
\item{data.df}{An ECIS dataframe}

\item{unit}{The unit to plot}

\item{frequency}{Frequency of interest, all modeled units have a frequency of 0}
}
\value{
A ggplot2 object
}
\description{
This function produces a 'spaghetti plot' of all replicates accross multiple experiments. This allows for the rapid identification of any erronious replicates in the context of the experimental variation observed
}
\examples{

ecis_plot_all(data.df, 'Rb', 0)


}
