% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot_all_timeslice}
\alias{ecis_plot_all_timeslice}
\title{Generate a bar graph of a particular value at a single time}
\usage{
ecis_plot_all_timeslice(data.df, unit, time)
}
\arguments{
\item{data.df}{An ECIS data set}

\item{unit}{Unit to plot}

\item{time}{Time at which the cross section should be taken}
}
\value{
ggplot2 object containing the dataset
}
\description{
Generate a bar graph of a particular value at a single time
}
\examples{

}
