% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_plot_badwell_scores}
\alias{ecis_plot_badwell_scores}
\title{Title}
\usage{
ecis_plot_badwell_scores(data, threshold = 0)
}
\arguments{
\item{data}{}

\item{threshold}{}
}
\value{

}
\description{
Title
}
\examples{
ecis_plot_badwell_scores(growth.df, 0.2)

}
