% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot_experiments}
\alias{ecis_plot_experiments}
\title{Title}
\usage{
ecis_plot_experiments(toplot.df, unit, frequency)
}
\arguments{
\item{unit}{}

\item{frequency}{}

\item{data.df}{}
}
\description{
Title
}
