% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_plot_frequency.R
\name{ecis_plot_frequency}
\alias{ecis_plot_frequency}
\title{Title}
\usage{
ecis_plot_frequency(
  data = growth.df,
  replication = "summary",
  unit = "R",
  time = 50,
  frequency = "raw",
  samplecontains = "",
  experiment = "",
  errorbars = Inf,
  errorcalc = "sem",
  alignkey = NULL,
  normtime = NULL,
  preprocessed = FALSE,
  continuouscontains = NULL,
  stripidentical = TRUE,
  alpha = 0.1
)
}
\arguments{
\item{data}{}

\item{unit}{}

\item{time}{}

\item{frequency}{}

\item{samplecontains}{}

\item{experiment}{}

\item{alignkey}{}

\item{normtime}{}

\item{preprocessed}{}

\item{continuouscontains}{}

\item{stripidentical}{}

\item{error}{}
}
\value{

}
\description{
Title
}
\examples{

ecis_plot_frequency(growth.df, replication = "summary", unit = "C", errorbars = Inf, samplecontains = "35,000")
ecis_plot_frequency(growth.df, replication = "experiments", unit = "C", errorbars = 1)
ecis_plot_frequency(growth.df, replication = "wells", unit = "C", errorbars = 1)
ecis_plot(growth.df)

ecis_plot_frequency(time = 100)

data = growth.df
replication = "wells"

}
