% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{ecis_plot_heatmap}
\alias{ecis_plot_heatmap}
\title{Title}
\usage{
ecis_plot_heatmap(data, time, unit, frequency, title = "Title")
}
\arguments{
\item{data}{}

\item{time}{}

\item{unit}{}

\item{frequency}{}
}
\value{

}
\description{
Title
}
\examples{

ecis_plot_heatmap(growth.df, 100, "Rb")
ecis_plot_heatmap(growth.df, 100, "R", 4000)

}
