% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_plot_spectra}
\alias{ecis_plot_spectra}
\title{Plot graphs accross a standard ECIS spectra}
\usage{
ecis_plot_spectra(data, variable, ...)
}
\arguments{
\item{data}{A standard ECIS data frame}

\item{variable}{Variable to plot. Can't be a modeled variable (use ecis_plot_model instead)}

\item{...}{Other arguements may be included that will be passed through to ecis_plot}
}
\value{
A matrix of ggplot2 objects
}
\description{
Plot graphs accross a standard ECIS spectra
}
\examples{

ecis_plot_spectra(growth.df, 'R', replication = "summary")

}
