% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot_summary}
\alias{ecis_plot_summary}
\title{Title}
\usage{
ecis_plot_summary(toplot.df, unit, frequency)
}
\arguments{
\item{toplot.df}{}

\item{unit}{}

\item{frequency}{}
}
\description{
Title
}
