% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot_summary_timeslice}
\alias{ecis_plot_summary_timeslice}
\title{Title}
\usage{
ecis_plot_summary_timeslice(data.df, unit, time)
}
\arguments{
\item{data.df}{A standard ECIS dataset}

\item{unit}{The unit to plot}

\item{time}{Time to plot a slice of}
}
\value{

}
\description{
Title
}
