% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_plotmodel}
\alias{ecis_plotmodel}
\title{Plot all values generated by the ECIS model}
\usage{
ecis_plotmodel(alldata.df)
}
\arguments{
\item{alldata.df}{An ECIS data frame}
}
\value{
A matrix containing graphs of all the variables generated by the ECIS model.
}
\description{
This is a pre-generated piece of code that automaticaly plots all the variables generated by the ECIS model against time. Resistance at 4000 Hz is also included as a sensible sanity check of the model.
}
\examples{

ecis_plotmodel(data.df)

}
