% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_plotspectra}
\alias{ecis_plotspectra}
\title{Plot graphs accross a standard ECIS spectra}
\usage{
ecis_plotspectra(data, variable)
}
\arguments{
\item{data}{A standard ECIS data frame}

\item{variable}{Variable to plot. Can't be a modeled variable.}
}
\value{
A matrix of ggplot2 objects
}
\description{
Plot graphs accross a standard ECIS spectra
}
\examples{

ecis_plotspectra(growth.df, 'R')

}
