% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plotvariable}
\alias{ecis_plotvariable}
\title{Title}
\usage{
ecis_plotvariable(data.df, unit, frequency)
}
\arguments{
\item{data.df}{}

\item{unit}{}

\item{frequency}{}
}
\description{
Title
}
\examples{

}
