% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_polish_plot}
\alias{ecis_polish_plot}
\title{Post-processing on ECIS plots to make them a bit more pretty}
\usage{
ecis_polish_plot(plot)
}
\arguments{
\item{plot}{The plot to format}
}
\value{
A standardised ggplot2 object
}
\description{
Post-processing on ECIS plots to make them a bit more pretty
}
\examples{
# Automatically applied to ggplot, should not be required externaly

}
