% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{ecis_prism}
\alias{ecis_prism}
\title{Export a prism-compatable data set}
\usage{
ecis_prism(data.df, unit, frequency = 0)
}
\arguments{
\item{data.df}{ECIS dataframe}

\item{unit}{Unit of data to export}

\item{frequency}{Frequency of data requored, modeled data defaults to 0}
}
\value{
A data frame that can be copied and pasted into prism
}
\description{
Export a prism-compatable data set
}
\examples{
ecis_prism(data.df, 'Rb')

}
