% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{ecis_reconstitute_sample}
\alias{ecis_reconstitute_sample}
\title{Reconstitute incorrectly formed sample field}
\usage{
ecis_reconstitute_sample(string)
}
\arguments{
\item{string}{The string to be converted to a continuous vector}
}
\value{
A CSV separated list of continuous variables
}
\description{
Uses string processing to guess what might be in an incorrectly formed string field to try and save things. It is generally better to go back and fix the import file than use this function. Legacy only as this is a very heavy function as it does tonnes of string maniuplations.
}
\examples{

ecis_reconstitute_sample("1nm cheese + 1nm cars")
ecis_reconstitute_sample("   5,000.939 nM Oranges")
ecis_reconstitute_sample("35,000 cells")

}
