% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_remove_metadata}
\alias{ecis_remove_metadata}
\title{Remove all non-core ECIS data from a data frame}
\usage{
ecis_remove_metadata(data.df, subset = "all")
}
\arguments{
\item{data.df}{}

\item{subset}{}
}
\description{
Usefull if you want to to further data manipluation, without having to worry about tracking multiple, unknown columns.
}
\examples{
exploded.df = ecis_explode(growth.df)
}
