% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_resample}
\alias{ecis_resample}
\title{Resample ECIS data onto a constant time base}
\usage{
ecis_resample(data.df, by, from = -Inf, to = Inf)
}
\arguments{
\item{by}{The frequency at which to resample}

\item{from}{The value at which to start resampling}

\item{to}{The max value to resample}

\item{data}{A standard ECIS data frame}
}
\value{
An ECIS dataset with re-located time points
}
\description{
This will currently over-sample data, use with care
}
\examples{

ecis_resample(growth.df, 8, 50 ,100)

}
