% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_standardise_wells}
\alias{ecis_standardise_wells}
\title{Standardise well names accross import types}
\usage{
ecis_standardise_wells(well)
}
\arguments{
\item{well}{The well to be standardised}
}
\value{
Standardised well names
}
\description{
Replaces A1 in strings with A01. Important for importing ABP files which may use either notation. Returns NA if the string could not be normalised, which can be configured to throw a warning in import code.
}
\examples{
ecis_standardise_wells('A01')
ecis_standardise_wells('A 1')
ecis_standardise_wells('tortoise') # Non-standardisable becomes NA
ecis_standardise_wells(growth.df$Well)

}
