% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_strip_badwells}
\alias{ecis_strip_badwells}
\title{Automatically strip badly connected wells from an ECIS dataset}
\usage{
ecis_strip_badwells(data.df, threshold = 5, frequency = 4000,
  unit = "R")
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{threshold}{How stringent to be in excluding wells. Higher is less stringent. Default is 5.}

\item{frequency}{Frequency to run numbers on, default is 4000}

\item{unit}{Unit to use in detection, default is R}
}
\value{
An ECIS dataframe, minus the offending wells. Will also print which wells have been excluded for manual review.
}
\description{
Automatically strip badly connected wells from an ECIS dataset
}
\examples{
ecis_strip_badwells (growth.df)


}
