% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{ecis_subset}
\alias{ecis_subset}
\title{Title}
\usage{
ecis_subset(data.df, time = Inf, unit = "", frequency = Inf,
  samplesubset = "", experiment = "")
}
\arguments{
\item{data.df}{}

\item{time}{}

\item{unit}{}

\item{frequency}{}

\item{samplesubset}{}
}
\description{
Title
}
\examples{
ecis_subset(data.df, time = 4.37, frequency = 4000, unit = "R", samplesubset = "05,000", experiment = "2")

}
