% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_methods.R
\name{ecis_summarise}
\alias{ecis_summarise}
\title{Summarise ECIS datasets from a single experiment}
\usage{
ecis_summarise(data.df)
}
\arguments{
\item{data.df}{An ECIS dataset in standard format}
}
\value{
An ECIS dataset supplimented with summary statistics
}
\description{
Creates and ECIS dataset that has had all samples of the same type averaged together. Assumes that each sample is independent, IE that this function has already been run on individual experiments
}
\examples{

ecis_summarise(growth.df)

}
