% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{ecis_test_exploded}
\alias{ecis_test_exploded}
\title{Test if an ECIS data frame is exploded}
\usage{
ecis_test_exploded(data)
}
\arguments{
\item{data}{the data frame to test}
}
\value{
True (exploded) or false (not exploded)
}
\description{
Test if an ECIS data frame is exploded
}
\examples{
ecis_test_exploded(growth.df)
imploded = ecis_implode(growth.df)
ecis_test_exploded(imploded)

}
