% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tests.R
\name{ecis_test_integrity}
\alias{ecis_test_integrity}
\title{Test the integrity of an ECIS dataframe}
\usage{
ecis_test_integrity(data.df, verbose = FALSE)
}
\arguments{
\item{data.df}{An ECIS dataframe, subset if needed}

\item{verbose}{How verbose to be. FALSE returns a dataset, TRUE returns a full verbose output of each test}
}
\value{
A whole slew of tests
}
\description{
This function will run a whole suite of tests on an ECIS dataframe, to check that it is both well designed for statistical analysis as well as technically intact
}
\examples{
# Prep a known good compliment of data
data = ecis_subset(growth.df, time = 100, unit = "R", frequency = 4000)

# Test a full compliment of data passes
ecis_test_integrity(data)

ecis_test_integrity(growth.df)
# Test function picks up unbalanced replicate #'s
data2 = ecis_exclude(data, well = c("A1", "B1", "C1"))
ecis_test_integrity(data2)
# Test function picks up missing pairs (in this case due to Rb not being established)
data3 = ecis_subset(growth.df, time = 100, unit = "Rb")
ecis_test_integrity(data3)

}
