% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tests.R
\name{ecis_test_summary_level}
\alias{ecis_test_summary_level}
\title{Test the summary level of an ECIS dataframe}
\usage{
ecis_test_summary_level(data.df)
}
\arguments{
\item{data.df}{A standard ECIS dataframe}
}
\value{
A summarised standard ECIS dataframe
}
\description{
Test the summary level of an ECIS dataframe
}
\examples{
experiment.df = ecis_summarise(growth.df, "experiment")
summary.df = ecis_summarise(growth.df, "summary")

ecis_test_summary_level(growth.df)
ecis_test_summary_level(experiment.df)
ecis_test_summary_level(summary.df)

}
