% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_titles}
\alias{ecis_titles}
\title{Title}
\usage{
ecis_titles(unit, frequency = 0)
}
\arguments{
\item{unit}{The unit to submit}

\item{frequency}{The frequency to submit}
}
\value{
An expression containing the correct data label for the unit
}
\description{
Title
}
\examples{

ecis_titles("Rb")

}
