% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_extension.R
\name{ecis_validate_file}
\alias{ecis_validate_file}
\title{Validate a file}
\usage{
ecis_validate_file(file_name, extension)
}
\arguments{
\item{file_name}{Character string specifying the file name}

\item{extension}{Character string containing a file extension that will be matched against the file type of file_name}
}
\value{
TRUE if it passes, FALSE if it does not. Also spits out warnings that will help the user correct the error
}
\description{
Validates if a file exists, and if it has the correct file extension.
}
\examples{
ecis_validate_file("R/AAA_TODOOO.R", "P")
ecis_validate_file("R/AAA_TODO.R", "P")
ecis_validate_file("R/AAA_TODO.R", "R")
}
