% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelligence.R
\name{excelligence_import}
\alias{excelligence_import}
\title{Import excelligence data files}
\usage{
excelligence_import(file, key)
}
\arguments{
\item{file}{Path to the excelligence file to be imported}

\item{key}{Optional, standard keyfile that will overwrite what is stored in the excelligence file}
}
\value{
A standard impedr data frame
}
\description{
Import excelligence data files
}
\examples{
filename = "inst/extdata/xcell.txt"
dataset = excelligence_import(filename)

ecis_plot(dataset, unit = "CI", frequency = 0 , replication = "experiments", time = c(170,220) , samplecontains = "Nerifollin", normtime = 185)

}
