% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getmode}
\alias{getmode}
\title{Calculate mode of a dataset}
\usage{
getmode(v)
}
\arguments{
\item{v}{Vector of numeric data to find the mode of}
}
\value{
The mode of the vector
}
\description{
Calculate mode of a dataset
}
\examples{
getmode(c(1,3,3,4,7))
}
