% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Title}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{}

\item{ncol}{}

\item{nrow}{}

\item{position}{}
}
\description{
Title
}
