% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{growth.df}
\alias{growth.df}
\title{Example data set from an ECIS growth experiment}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Experiment}{Date the experiment was conducted}
  \item{Frequency}{Frequency of colleciton, modeled units have a frequency of 0}
  \item{Sample}{Name of the sample located in that well}
  \item{Time}{Capture time}
  \item{Unit}{Unit that the value is measured in}
  \item{Value}{Numerical value in units}
  \item{Well}{Well of plate that was measured}
  \item{cells}{Number of cells added to each well}
  \item{line}{The cell line used in this experiment (all HCMEC/D3)}
  ...
}}
\source{
James Hucklesby 2019
}
\usage{
growth.df
}
\description{
A dataset containing growth curves from a variety of experiments
}
\keyword{datasets}
