% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xcelligence.R
\name{import_mdb}
\alias{import_mdb}
\title{Internal Function to import MDB files from access}
\usage{
import_mdb(file, table, password)
}
\arguments{
\item{file}{Path to the file to import}

\item{table}{Name of the acess table to be imported}
}
\value{
A table, as outlined in the Access database
}
\description{
xCELLigence files are stored in a proprietary Microsoft database format. This funciton launches a local copy of 32 bit R, then connects to Access and pulls the table through. This is in a separate function to the main file, as multiple sessions are needed to pull out various tables and this is the cleanest way to achieve that.
}
\details{
This funciton will only run if Microsoft Access is installed on the PC
}
\examples{
# This file is used in import_cellzScope exclisivley so can be tested with it

}
\keyword{internal}
