% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xcelligence.R
\name{import_xcelligence}
\alias{import_xcelligence}
\title{Import an xcelligence file}
\usage{
import_xcelligence(file, key)
}
\arguments{
\item{file}{The file to import}

\item{key}{A keyfile to apply. Optional, as the xCELLigence internal definitions will be used if no file is specified}
}
\value{

}
\description{
Import an xcelligence file
}
\examples{
# Arguments to push through the function
file = "inst/extdata/xcell.plt"
key = "inst/extdata/xcell_lookup.csv"
xcell = import_xcelligence(file, key)

xcell = ecis_explode(xcell)

#ecis_plot(xcell, unit = "Z", frequency = "10000", replication = "experiments", continuouscontains = "ATP", normtime = 0)

#ecis_plot(xcell, unit = "CI", frequency = 10000, continuouscontains = "PDGF", replication = "wells")

 
}
