% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_updates.R
\name{install_git_requirements}
\alias{install_git_requirements}
\title{Title}
\usage{
install_git_requirements()
}
\value{
Installs all the git requisites for the package. Not available via CRAN.
}
\description{
Title
}
\examples{

install_git_requirements()

}
