% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{locked_median}
\alias{locked_median}
\title{Title}
\usage{
locked_median(data_vector)
}
\arguments{
\item{data_vector}{A vector of data to calculate the locked median of}
}
\description{
Title
}
\keyword{internal}
