% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_shiny.R
\name{mc_tribble}
\alias{mc_tribble}
\title{Make a tribble from underlying R code}
\usage{
mc_tribble(toprint)
}
\arguments{
\item{toprint}{the tribble to print}
}
\value{
A copy and pasteable tribble
}
\description{
Make a tribble from underlying R code
}
\examples{
mc_tribble(indf = growth.df[1:4,])


}
