% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr2-package.R
\docType{data}
\name{rbgrowth.df}
\alias{rbgrowth.df}
\title{Vascr growth data which is not re sampled}
\format{
A tibble with 346370 rows and 9 variables:
\describe{
  \item{Time}{The time at which the measurement was taken (hours)}
  \item{Unit}{The unit the measurement was taken in}
  \item{Well}{The well in which the measurement was taken}
  \item{Value}{The value of the measurement taken}
  \item{Frequency}{The frequency at which data was collected}
  \item{Experiment}{Name of the experiment}
  \item{Instrument}{The instrument data was collected on}
  \item{SampleID}{The numerical ID of the sample}
  \item{Sample}{The name of the treatment applied to the dataset}
}
}
\source{
Hucklesby 2020
}
\usage{
rbgrowth.df
}
\description{
A dataset containing the growth curves of hCMEC/D3 cell lines seeded at various densities. 
This is a subset of 'growth.df', with only Rb data for faster processing.
}
\keyword{internal}
