% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr2-package.R
\name{\%>\%}
\alias{\%>\%}
\alias{tribble}
\alias{glue}
\title{re-export magrittr pipe operator}
\value{
Various outputs, please see documentation for source package

description Used to easily make pipe operator available

A tibble, made available for easy data entry manually

A string with glue substitution completed
}
\description{
re-export magrittr pipe operator

re-export tribble

re-export glue
}
