% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_updates.R
\name{reinstall_updates}
\alias{reinstall_updates}
\title{Reinstalls the package from GitHub. Temporary.}
\usage{
reinstall_updates()
}
\value{
Installs all the git requisites for the package. Not available via CRAN.
}
\description{
Reinstalls the package from GitHub. Temporary.
}
\examples{

#install_git_requirements()

}
