% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_cols_if_exists}
\alias{remove_cols_if_exists}
\title{Remove columns if they exist in a dataset, otherwise do nothing}
\usage{
remove_cols_if_exists(data.df, cols)
}
\arguments{
\item{data.df}{The dataset to remove}

\item{cols}{The name, or names, of cols to remove}
}
\value{
The truncated data set
}
\description{
Remove columns if they exist in a dataset, otherwise do nothing
}
\examples{
remove_cols_if_exists(data.df, c("Unit", "Sample", "Donkey"))
}
