% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{retitle}
\alias{retitle}
\title{Retitle}
\usage{
retitle(df)
}
\arguments{
\item{df}{A data frame containing the desired values in the top row}
}
\value{
A dataframe where the top row has been converted to titles.
}
\description{
Recapitulation of the funciton in tidyR, allows the re-titling of a data frame from the top row of a dataset. Used in import funcitons to set titles from the content of ABP files. For internal use only.
}
