% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr2-package.R
\docType{data}
\name{small_growth.df}
\alias{small_growth.df}
\title{Small growth data set for testing}
\format{
A tibble with 346370 rows and 9 variables:
\describe{
  \item{Time}{The time at which the measurement was taken (hours)}
  \item{Unit}{The unit the measurement was taken in}
  \item{Well}{The well in which the measurement was taken}
  \item{Value}{The value of the measurement taken}
  \item{Frequency}{The frequency at which data was collected}
  \item{Experiment}{Name of the experiment}
  \item{Instrument}{The instrument data was collected on}
  \item{SampleID}{The numerical ID of the sample}
  \item{Sample}{The name of the treatment applied to the dataset}
}
}
\source{
Hucklesby 2020
}
\usage{
small_growth.df
}
\description{
Small growth data set for testing
}
\keyword{internal}
