% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr2-package.R
\docType{package}
\name{vascr-package}
\alias{vascr}
\alias{vascr-package}
\title{vascr: Process Biological Impedance Sensing Data}
\description{
Process complex impedance sensing datasets, including those generated by ECIS, xCELLigence and cellZscope instruments. Data can be imported to a standardised tidy format and then plotted. Support for conducting and plotting the outputs of ANOVA (with appropriate tests of statistical assumptions) and cross-correlation analysis. For data processed using this package see Hucklesby et al. (2020) \doi{10.3390/bios11050159}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesHucklesby/vascr}
  \item Report bugs at \url{https://github.com/JamesHucklesby/vascr/issues}
}

}
\author{
\strong{Maintainer}: James JW Hucklesby \email{james.hucklesby@auckland.ac.nz} (\href{https://orcid.org/0000-0003-1591-6823}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Charles Unsworth \email{c.unsworth@auckland.ac.nz} (\href{https://orcid.org/0000-0002-9153-5232}{ORCID})
  \item E Scott Graham \email{s.graham@auckland.ac.nz} (\href{https://orcid.org/0000-0002-2925-9126}{ORCID})
  \item Catherine E angel \email{c.angel@auckland.ac.nz} (\href{https://orcid.org/0000-0003-1737-8539}{ORCID})
}

}
\keyword{internal}
