% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_96_well_names}
\alias{vascr_96_well_names}
\title{All the well names of a 96 well plate}
\usage{
vascr_96_well_names()
}
\value{
Vector containing all wells of a 96 well plate
}
\description{
All the well names of a 96 well plate
}
\examples{
#vascr_96_well_names()

}
