% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_all_null}
\alias{vascr_all_null}
\title{Check if all arguements are null}
\usage{
vascr_all_null(...)
}
\arguments{
\item{...}{List of items, all of which may be null}
}
\value{
Boolean
}
\description{
Check if all arguements are null
}
\examples{
# vascr_all_null(NULL, TRUE)
}
\keyword{internal}
