% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_append_unit}
\alias{vascr_append_unit}
\title{Append the unit containing string from a second vector to a non-unit containing string}
\usage{
vascr_append_unit(strippedcol, allcols)
}
\arguments{
\item{strippedcol}{The column without a string present}

\item{allcols}{The full colunm names to search}
}
\value{
The appended unit
}
\description{
Append the unit containing string from a second vector to a non-unit containing string
}
\examples{
vascr_append_unit("ATP", xcell)
}
