% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_naming.R
\name{vascr_apply_map}
\alias{vascr_apply_map}
\title{Apply a map to a vascr dataset}
\usage{
vascr_apply_map(data.df, map)
}
\arguments{
\item{data.df}{the dataset to apply to}

\item{map}{the dataset to apply}
}
\value{
a named vascr dataset
}
\description{
Apply a map to a vascr dataset
}
\examples{
lookup = system.file('extdata/instruments/eciskey.csv', package = 'vascr')
vascr_apply_map(data.df = growth.df, map = lookup)

vascr_apply_map(growth.df \%>\% vascr_subset(well = c("A1")), lookup)

}
