% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_import_ecis.R
\name{vascr_assign_samples}
\alias{vascr_assign_samples}
\title{Internal function to assign sample names to a data frame}
\usage{
vascr_assign_samples(data, sampledefine)
}
\arguments{
\item{data}{an ECIS data frame}

\item{sampledefine}{path to the file containing the sample definitions}
}
\value{
A standard ECIS data frame, annotated
}
\description{
Internal function to assign sample names to a data frame
}
\examples{
# This function is baked into ecis_import and it's parts

}
\keyword{internal}
