% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_carry_down_names}
\alias{vascr_carry_down_names}
\title{Carry down names in selected columns}
\usage{
vascr_carry_down_names(data.df, cols_to_carry = NULL, remove_blank = TRUE)
}
\arguments{
\item{data.df}{The data frame to carry down columns in}

\item{cols_to_carry}{The columns to carry down, default is the whole lot}

\item{remove_blank}{Should blank names (starting with _, 0_ or NA_) be removed}
}
\value{
A data frame with columns carried down
}
\description{
Carry down names in selected columns
}
\examples{
# Used as part of various import functions

}
\keyword{internal}
