% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_ccf_pairs}
\alias{vascr_ccf_pairs}
\title{Title}
\usage{
vascr_ccf_pairs(data.df, reference = NULL, comparator = NULL)
}
\arguments{
\item{data.df}{}

\item{type}{}
}
\value{

}
\description{
Title
}
\examples{
growthrb = vascr_subset(growth.df, unit = "Rb", time = c(5,150))
deltadata = vascr_summarise_change(growthrb)
ret = vascr_ccf_pairs(deltadata)
ret = vascr_ccf_pairs(deltadata, reference = "internal")
ret = vascr_ccf_pairs(deltadata, reference = "[ 1 : Experiment 1 | 35000 ]")
ret = vascr_ccf_pairs(deltadata, reference = "[ 1 : Experiment 1 | 25000 ]", comparator = "[ 1 : Experiment 1 | 30000 ]")


}
