% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_naming.R
\name{vascr_change_name}
\alias{vascr_change_name}
\title{Edit a sample name in a vascr dataframe}
\usage{
vascr_change_name(data.df, sampleid, new_name = "")
}
\arguments{
\item{data.df}{The data set to edit}

\item{to_remove}{The sample to remove}

\item{to_add}{The sample to replace with}
}
\value{
An edited vascr dataset
}
\description{
Edit a sample name in a vascr dataframe
}
\examples{
changed = vascr_change_name(growth.df, 1 , "Sample One")
}
