% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_col_exists}
\alias{vascr_col_exists}
\title{Title}
\usage{
vascr_col_exists(col, data.df)
}
\arguments{
\item{col}{}

\item{data.df}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_col_exists("Well", data.df)
}
