% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_collapse_hash}
\alias{vascr_collapse_hash}
\title{Collapse multiple hashes in a string down to a single hash}
\usage{
vascr_collapse_hash(string)
}
\arguments{
\item{string}{The string to collapse the hashes in}
}
\value{
A string with hashes collapsed
}
\description{
Collapse multiple hashes in a string down to a single hash
}
\examples{
#vascr_collapse_hash("###cat###andthe##hat####")


}
\keyword{internal}
