% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{vascr_cols}
\alias{vascr_cols}
\title{Return the ECIS cols used in this package}
\usage{
vascr_cols(data, set = "core")
}
\arguments{
\item{data}{The dataset to work off. Only required if non-standard cols are requested}

\item{set}{The set of columns to request. Default is core.}
}
\value{
A vector of the columns requested
}
\description{
Can return either the core set of columns required for an ECIS package, the continous or categorical variables, or the exploded variables. Set will return the lot as a list.
}
\examples{

#vascr_cols()
#vascr_cols(growth.df, set = "exploded")
#vascr_cols(growth.df, set = "continuous")
#vascr_cols(growth.df, set = "categorical")
#vascr_cols(growth.df, set = "set")
#vascr_cols(growth.df, set = "is_continuous")


}
\keyword{internal}
