% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_current_frequency}
\alias{vascr_current_frequency}
\title{Current aquisition rate}
\usage{
vascr_current_frequency(data.df)
}
\arguments{
\item{data.df}{The dataframe to compute the current data aquisition frequency of}
}
\value{
The current aquisition rate of the data frame
}
\description{
Current aquisition rate
}
\examples{

#vascr_current_frequency(growth.df)


}
