% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_default}
\alias{vascr_default}
\title{Look up the default variable values for all supported instruments}
\usage{
vascr_default(data)
}
\arguments{
\item{data}{The dataset to derrive the instrument type from}
}
\value{
The default frequency and unit in a list
}
\description{
Central source of lookup values for default graph settings for various instruments.
}
\examples{
# vascr_default(growth.df)

}
\keyword{internal}
