% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_quality.R
\name{vascr_detect_deviation}
\alias{vascr_detect_deviation}
\title{Detect the deviation between wells and the median well}
\usage{
vascr_detect_deviation(data, deviation = 0, frequency = 4000, unit = "R")
}
\arguments{
\item{data}{A vascr dataset to process}

\item{deviation}{The minimum deviation to return}

\item{frequency}{The frequency to run the calculations on. Only one frequency should be selected, default is 4000.}

\item{unit}{The unit to run calculations on. Only one unit should be selected, and generally this should be a raw frequency. Default R.}
}
\value{
A vascr dataset containing the deviaiton at each time point
}
\description{
This function detects the proportion deviation between the median of each sample and each well in that sample for each time point. This will return zero for the median wel at each time point, however which well is the median may change between time points.
}
\details{
This function is not as diagnostic or sensitive compared to vascr_detect_max_deviation but is usefull if you need to find out when and where a well differs from the remainder of the triplicate.
}
\examples{
example = vascr_detect_deviation(growth.df)

}
