% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_find.R
\name{vascr_detect_level}
\alias{vascr_detect_level}
\title{Check the level of a vascr data frame}
\usage{
vascr_detect_level(data)
}
\arguments{
\item{data}{The data frame to analyse}
}
\value{
The level of the dataset analysed
}
\description{
Check the level of a vascr data frame
}
\examples{
#vascr_detect_level(growth.df)
#vascr_detect_level(vascr_summarise(growth.df, level = "experiments"))
#vascr_detect_level(vascr_summarise(growth.df, level = "summary"))
}
