% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_quality.R
\name{vascr_detect_max_deviation}
\alias{vascr_detect_max_deviation}
\title{Automatically detect erronious wells in a vascr dataset}
\usage{
vascr_detect_max_deviation(
  data.df,
  max_deviation = 0,
  frequency = 4000,
  unit = "R"
)
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{max_deviation}{Maximum deviation that is acceptable, lower is more stringent}

\item{frequency}{Frequency to run numbers on, default is 4000}

\item{unit}{Unit to use in detection, default is R}
}
\value{
A tibble containing the offending wells, which experiment they are from and the score they were removed with
}
\description{
As with all high-throughput measurements, replicaes can behave badly. This dataset automatically detects these values.
}
\details{
This function is sensitive to two factors, wells with large differences between replicates and sudden spikes in datasets caused by a transient loss of connection. Max_Deviation is calculated by determining the fractional difference between any well and the median well for that sample:experiment combination at each individual timepoint. This will identify wells that sit well away from the centre of the dataset, as median is not affected by large outliers. Secondly we select the largest of the values calculated for each time point and show this as the score. This makes the score highly sensitive to transient spikes in the dataset, that would be dulled by averaging the dataset as a whole.

This function can be applied to a whole dataset. Selecting smaller areas does not make the detection more sensitive, but can be used to exclude particular wells. Various units can be screened, though raw units are often more appropriate for diagnosing technical issues.

A dataframe is returned with a score for each well, fully labeled minus Time and Value as these are inappropriate for a summarised unit.
}
\examples{
# Return the results as a table
# vascr_detect_max_deviation(growth.df)
# Then return a graphical representaiton
#vascr_plot_deviation(growth.df)
}
\keyword{internal}
