% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_detect_normal}
\alias{vascr_detect_normal}
\title{Detect if an ECIS dataset has been normalised}
\usage{
vascr_detect_normal(data.df)
}
\arguments{
\item{data.df}{an ECIS dataset}
}
\value{
The time the data was normalised to, or FALSE if not normalised
}
\description{
Detect if an ECIS dataset has been normalised
}
\examples{
#growth.df$Instrument = "ECIS"
#standard = growth.df
#normal = vascr_normalise(growth.df, 100)
#vascr_detect_normal(standard)
#vascr_detect_normal(normal)
}
