% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_continuous.R
\name{vascr_detect_vehicle_control_wells}
\alias{vascr_detect_vehicle_control_wells}
\title{Detect the wells R thinks contains vehicles}
\usage{
vascr_detect_vehicle_control_wells(data, force_replace = FALSE)
}
\arguments{
\item{data}{The datset to search}

\item{force_replace}{Should existing IsVehicleControl wells be replaced (if present)?}
}
\value{
A vector of vehicle control wells
}
\description{
Convenience function. This first runs vascr_detect_vehicle on the dataset, then provides a summary table of which wells are associated with which vehicle. Provides a usefull check that vehicle detection has been preformed correctly.
}
\examples{
#vascr_detect_vehicle_control_wells(xcell)

}
\keyword{internal}
