% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_df_title}
\alias{vascr_df_title}
\title{Title}
\usage{
vascr_df_title(data.df)
}
\arguments{
\item{data.df}{}
}
\value{

}
\description{
Title
}
\examples{
data.df = growth.df \%>\% vascr_subset(unit = "Rb") \%>\% 
vascr_normalise(normtime = 100)


}
