% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_naming.R
\name{vascr_edit_name}
\alias{vascr_edit_name}
\title{Edit a sample name in a vascr dataframe}
\usage{
vascr_edit_name(data.df, to_remove, to_add = "")
}
\arguments{
\item{data.df}{The data set to edit}

\item{to_remove}{The sample to remove}

\item{to_add}{The sample to replace with}
}
\value{
An edited vascr dataset
}
\description{
Edit a sample name in a vascr dataframe
}
\examples{
vascr_edit_name(growth.df,"HCMEC D3", "HCMEC/D3")
}
