% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_exclude}
\alias{vascr_exclude}
\title{#' Fix typographical errors in an ECIS dataframe
#' 
#' This function will go through all the sample and experiment names and replace all the values that are incorrect. Usefull for fixing up typographical errors, or places where you have named things differently inadvertently.
#'
#' @param data.df the data to correct
#' @param incorrect the error to be corrected
#' @param correct the corrrect string
#' @param limit can be set to "Sample or "Experiment" to limit the replacement to a single variable
#'
#' @return the repared dataframe
#' @export
#'
#' @examples
#' 
#' #vascr_fix_error(growth.df, "cells", "cells plated")
#' 
vascr_fix_error = function (data.df, incorrect, correct, limit = "None")
{
  
  if (limit == "None" || limit == "Sample")
  {
  data.df$Sample = gsub(incorrect, correct, data.df$Sample)
  }
  
  if (limit == "None" || limit == "Experiment")
  {
  data.df$Experiment = gsub(incorrect, correct, data.df$Experiment)
  }
  
  return(data.df)
}
Exclude erronious data from an ECIS dataframe}
\usage{
vascr_exclude(
  data.df,
  samples = FALSE,
  wells = FALSE,
  experiments = FALSE,
  times = FALSE,
  values = FALSE,
  vars = FALSE,
  vals = FALSE,
  vs = FALSE
)
}
\arguments{
\item{data.df}{The source dataset}

\item{samples}{The sample(s) to exclude}

\item{wells}{The well(s) to exclude}

\item{experiments}{The experiment(s) to exclude}

\item{times}{The time(s) to exclude}

\item{values}{The value(s) to exclude}

\item{vars}{The isolated variables to exclude}

\item{vals}{The isolated values to exclude}

\item{vs}{The isolated variables-unit pairs to exclude}
}
\value{
The altered dataset
}
\description{
#' Fix typographical errors in an ECIS dataframe
#' 
#' This function will go through all the sample and experiment names and replace all the values that are incorrect. Usefull for fixing up typographical errors, or places where you have named things differently inadvertently.
#'
#' @param data.df the data to correct
#' @param incorrect the error to be corrected
#' @param correct the corrrect string
#' @param limit can be set to "Sample or "Experiment" to limit the replacement to a single variable
#'
#' @return the repared dataframe
#' @export
#'
#' @examples
#' 
#' #vascr_fix_error(growth.df, "cells", "cells plated")
#' 
vascr_fix_error = function (data.df, incorrect, correct, limit = "None")
{
  
  if (limit == "None" || limit == "Sample")
  {
  data.df$Sample = gsub(incorrect, correct, data.df$Sample)
  }
  
  if (limit == "None" || limit == "Experiment")
  {
  data.df$Experiment = gsub(incorrect, correct, data.df$Experiment)
  }
  
  return(data.df)
}
Exclude erronious data from an ECIS dataframe
}
\examples{

#unique(growth.df$Sample)
#excludedgrowth.df = vascr_exclude(growth.df, samples = c("35,000 cells", "0 cells"))
#unique(excludedgrowth.df$Sample)

#unique(growth.df$Well)
#excludedgrowth.df = vascr_exclude(growth.df, wells = c("A1", "B1", "C1"))
#unique(excludedgrowth.df$Well)

#unique(growth.df$Experiment)
#excludedgrowth.df = vascr_exclude(growth.df, experiment = c(1,2))
#unique(excludedgrowth.df$Experiment)

}
