% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{vascr_explode}
\alias{vascr_explode}
\title{Explode out continuous data from Sample}
\usage{
vascr_explode(data)
}
\arguments{
\item{data}{Standard ECIS data frame to explode}
}
\value{
The data frame to return
}
\description{
Explode out continuous data from Sample
}
\examples{
# growth.df$Instrument ="ECIS"


}
