% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_explode_wells}
\alias{vascr_explode_wells}
\title{Explode the wells in a VASCR dataset}
\usage{
vascr_explode_wells(data, separate_rows = FALSE)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Explode the wells in a VASCR dataset
}
\examples{
vascr_explode_wells(growth.df)

data = vascr_summarise(growth.df)
vascr_explode_wells(data)
vascr_explode_wells(data, separate_rows = TRUE)

}
