% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_export.R
\name{vascr_export}
\alias{vascr_export}
\title{Title}
\usage{
vascr_export(data.df, filepath = NULL)
}
\arguments{
\item{data.df}{}

\item{filepath}{}
}
\value{
A dataframe in prism format, or writes to file if filepath specified
}
\description{
Title
}
\examples{
filepath = tempfile("test_export", fileext = ".xlsx")
vascr_export(data.df, filepath)

}
