% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{vascr_export_prism}
\alias{vascr_export_prism}
\title{Export files to prism en-masse, and save them in a folder}
\usage{
vascr_export_prism(data, filename, units, frequency = 0)
}
\arguments{
\item{data}{The data frame to export from}

\item{filename}{The file to export to. If multiple units are selected, this will also be the name of the folder that a file for each unit is saved into.}

\item{units}{The unit(s) to export. Will accept a string or vector. Stating "all" will export every frequency and unit combination in the dataset.}

\item{frequency}{The frequency, if needed and exporting a single unit.}
}
\value{
Multiple CSV files, in the best format for GraphPad Prism
}
\description{
This will generate a folder full of CSV files that can be imported to prism, should you want to plot the data there.
}
\examples{
# Export a massive dataset. Not run as an example because it makes a mess.

# vascr_export_prism(growth.df, "PRISMTEST5", "all")

}
