% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_factorise_and_sort}
\alias{vascr_factorise_and_sort}
\title{Factorise and sort a the column of a data frame for plotting}
\usage{
vascr_factorise_and_sort(data, sortkeyincreasing = TRUE)
}
\arguments{
\item{data}{A vector to sort and factorise. This uses stringr's numeric =TRUE to line up all numbers in order correctly}

\item{sortkeyincreasing}{Should samples be returned increasing or decreasing. Default increasing.}
}
\description{
Factorise and sort a the column of a data frame for plotting
}
