% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_find_changing_cols}
\alias{vascr_find_changing_cols}
\title{Find which R columns change accross the dataset}
\usage{
vascr_find_changing_cols(data.df)
}
\arguments{
\item{data.df}{The dataset to analyse}
}
\value{
A vector of the column names that change in the dataset
}
\description{
Find which R columns change accross the dataset
}
\examples{
# vascr_find_changing_cols(data.df)
}
\keyword{internal}
