% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_colname}
\alias{vascr_find_colname}
\title{Find exploded column names in teh dataset}
\usage{
vascr_find_colname(data.df, colname = NULL)
}
\arguments{
\item{data.df}{The dataset to analyse}

\item{colname}{The column to look for}
}
\value{
The columns that match
}
\description{
Find exploded column names in teh dataset
}
\examples{
#vascr_find_colname(data.df, "ADP")
#vascr_find_colname(data.df, colname = "HCMVEC")

}
\keyword{internal}
