% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_experiment}
\alias{vascr_find_experiment}
\title{Find an experiment in a vascr dataset}
\usage{
vascr_find_experiment(data.df, experiment)
}
\arguments{
\item{data.df}{The dataset to find the experiment in}

\item{experiment}{The experiment to find. Either a name or a number from when the experiment is combined can be used.}
}
\value{
A character string of the most closley matched experiment
}
\description{
Find an experiment in a vascr dataset
}
\examples{
# vascr_find_experiment(growth.df, 1)
# vascr_find_experiment(growth.df, "1 : Experiment 1")

}
\keyword{internal}
