% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_instrument}
\alias{vascr_find_instrument}
\title{Find an instrument known to vascr}
\usage{
vascr_find_instrument(data.df, instrument = NULL)
}
\arguments{
\item{data.df}{A vascr dataset to analyse}

\item{instrument}{The instrument to search for}
}
\value{
A vector of strings that match
}
\description{
Find an instrument known to vascr
}
\examples{
#vascr_find_instrument(growth.df, "Rb")
#vascr_find_instrument(growth.df, "cellZScope")
#vascr_find_instrument(growth.df, c("cellZscope", "ECIS"))

#vascr_find_instrument(growth.df, NULL)
}
\keyword{internal}
