% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_find.R,
%   R/vascr_find.sync-conflict-20260526-135829-YEPQG6Q.R
\name{vascr_find_level}
\alias{vascr_find_level}
\title{Check the level of a vascr data frame}
\usage{
vascr_find_level(data)

vascr_find_level(data)
}
\arguments{
\item{data}{The data frame to analyse}
}
\value{
The level of the dataset analysed

The level of the dataset analysed
}
\description{
Check the level of a vascr data frame

Check the level of a vascr data frame
}
\examples{
vascr_find_level(growth.df)
vascr_find_level(vascr_summarise(growth.df \%>\% vascr_subset(unit= "Rb"), level = "experiments"))
vascr_find_level(vascr_summarise(growth.df \%>\% vascr_subset(unit= "Rb"), level = "summary"))
vascr_find_level(growth.df)
vascr_find_level(vascr_summarise(growth.df \%>\% vascr_subset(unit= "Rb"), level = "experiments"))
vascr_find_level(vascr_summarise(growth.df \%>\% vascr_subset(unit= "Rb"), level = "summary"))
}
