% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_find.R,
%   R/vascr_find.sync-conflict-20260526-135829-YEPQG6Q.R
\name{vascr_find_metadata}
\alias{vascr_find_metadata}
\title{Print out the characteristics of the vascr data frame}
\usage{
vascr_find_metadata(data.df)

vascr_find_metadata(data.df)
}
\arguments{
\item{data.df}{the vascr data frame to interrogate}
}
\value{
prints out the parameters of the dataframe in question

prints out the parameters of the dataframe in question
}
\description{
Print out the characteristics of the vascr data frame

Print out the characteristics of the vascr data frame
}
\examples{
vascr_find_metadata(growth.df)

vascr_find_metadata(growth.df)

}
