% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_sample}
\alias{vascr_find_sample}
\title{Find vascr samples that meet the criteria}
\usage{
vascr_find_sample(
  data.df,
  sample_contains = NULL,
  sample_not_contains = NULL,
  sample_equal_to = NULL,
  variable_set = NULL,
  variable_unset = NULL,
  variable_equal_to = NULL,
  variable_not_equal_to = NULL,
  variable_greater_than = NULL,
  variable_less_than = NULL,
  include_vehicle = TRUE
)
}
\arguments{
\item{data.df}{The dataset to screen}

\item{sample_contains}{Returns that samples that contain a string. This is very broad.}

\item{sample_not_contains}{Include sampels that do not contain this value}

\item{sample_equal_to}{Return where the sample is an exact match}

\item{variable_set}{Returns where the variable is set}

\item{variable_unset}{Returns where the variable is unset}

\item{variable_equal_to}{Returns if the variable, is equal to the second unit in the vector}

\item{variable_not_equal_to}{Returns when the variable is not equal to the second in each pair}

\item{variable_greater_than}{Returns when the varaible first is greater than the second value.}

\item{variable_less_than}{Returns when the variable is less than the second value}

\item{include_vehicle}{Should the vehicle sample be included in the dataset}
}
\value{
A vector of samples that match the criterion
}
\description{
Search the samples, based on pairs of datasets. Pass to any variable names of variables paired with what is beng screened for. Using include_vehicle will respect variable
}
\examples{

# vascr_find_sample(xcell, variable_set = "ATP")
# vascr_find_sample(growth.df)

# vascr_find_sample(czs)

# data.df = vascr_subset(unifiedr, unit = "Rb")
# vascr_find_sample(data.df, variable_equal_to = c("HCMVEC", 20000))

}
\keyword{internal}
