% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_find.R
\name{vascr_find_sampleid_from_sample}
\alias{vascr_find_sampleid_from_sample}
\title{Title}
\usage{
vascr_find_sampleid_from_sample(data.df, sample)
}
\arguments{
\item{data.df}{}

\item{sampleID}{}
}
\description{
Title
}
\examples{
vascr_find_sampleid_from_sample(growth.df, "5,000_cells + HCMEC D3_line")

}
