% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_single_time}
\alias{vascr_find_single_time}
\title{Find a time aligned with a vascr dataset}
\usage{
vascr_find_single_time(data.df, time)
}
\arguments{
\item{data.df}{The dataset to align to}

\item{time}{The time to align}
}
\value{
A single time that aligns with the dataset
}
\description{
Find a time aligned with a vascr dataset
}
\examples{
vascr_find_time(data.df, 43.78)
}
