% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_find_time}
\alias{vascr_find_time}
\title{Align times}
\usage{
vascr_find_time(data.df, time)
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{time}{The time point that needs rounding}
}
\value{
A timepoint that exactly aligns with a measured datapoint
}
\description{
When running analyasis, you can only run stats on a timepoint that exists in the dataset. These are not always logical or easy to remember. This function rounds the number given to the nearest timepoint that is actually in the dataset.
}
\examples{
vascr_find_time(growth.df, 146.2)

}
