% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_unit}
\alias{vascr_find_unit}
\title{Find a valid unit in the dataset, and throw an error if the unit selected is not appropriate}
\usage{
vascr_find_unit(data.df, unit = NA)
}
\arguments{
\item{data.df}{The dataset to find the unit in}

\item{unit}{The unit to find}
}
\value{
A vector of units that have been identified
}
\description{
Find a valid unit in the dataset, and throw an error if the unit selected is not appropriate
}
\examples{
# vascr_find_unit(growth.df, "raw")
# vascr_find_unit(growth.df, "modeled")
# vascr_find_unit(growth.df, "all")
# vascr_find_unit(growth.df, "Cm")
# vascr_find_unit(growth.df, NA)
# vascr_find_unit(growth.df, unit = c("Ci", "Rb"))

}
\keyword{internal}
