% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_value}
\alias{vascr_find_value}
\title{Find values within particular paramaters}
\usage{
vascr_find_value(
  data.df,
  value_greater_than = NULL,
  value_less_than = NULL,
  value_equal_to = NULL
)
}
\arguments{
\item{data.df}{The dataset to analyse}

\item{value_greater_than}{Values will be returned if they're greater than this value}

\item{value_less_than}{Values will be returned if less than this value}

\item{value_equal_to}{Values will be returned if equal to this value}
}
\value{
A list of values that fit the criteria
}
\description{
Find values within particular paramaters
}
\examples{
vascr_find_value(data.df)

}
