% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_well}
\alias{vascr_find_well}
\title{Find a well in a local dataset}
\usage{
vascr_find_well(data.df, well)
}
\arguments{
\item{data.df}{The dataset to detect from}

\item{well}{The well to find}
}
\value{
The string of a valid well to return
}
\description{
Find a well in a local dataset
}
\examples{
# vascr_find_well(growth.df, "A1")
# vascr_find_well(growth.df, NULL)

}
\keyword{internal}
