% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_formula}
\alias{vascr_formula}
\title{Generate an anova formula}
\usage{
vascr_formula(priority)
}
\arguments{
\item{priority}{The priority to use}
}
\value{
An ANOVA formula
}
\description{
Generate an anova formula
}
\examples{
# vascr_formula(c("cells","well"))
}
