% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_full_implode}
\alias{vascr_full_implode}
\title{Title}
\usage{
vascr_full_implode(data.df, cols_to_implode = NULL)
}
\arguments{
\item{data.df}{}

\item{cols_to_implode}{}
}
\value{

}
\description{
Title
}
