% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_grab_legend}
\alias{vascr_grab_legend}
\title{Title}
\usage{
vascr_grab_legend(aplot, position = "left")
}
\arguments{
\item{aplot}{}
}
\value{

}
\description{
Title
}
\examples{
ledgend = grid.arrange(vascr_grab_legend(p1, position = "right"))
grid.arrange(vascr_grab_legend(p1, position = "top"))

ggplotGrob(ledgend)


}
