% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_import.R
\name{vascr_import}
\alias{vascr_import}
\title{Import impedance data to a vascr dataset}
\usage{
vascr_import(
  instrument,
  raw = NULL,
  model = NULL,
  key = NULL,
  experimentname = "NA",
  password = "RTCaDaTa"
)
}
\arguments{
\item{instrument}{The instrument to import from.}

\item{raw}{Path to the raw dataset. Ignored if not set}

\item{model}{Path to the modeled dataset. Ignored if not set}

\item{key}{Key file, containing data on which well contains what data}

\item{experimentname}{Name to assign to the experiment}
}
\value{
A full vascr dataset
}
\description{
Import impedance data to a vascr dataset
}
\examples{
# ECIS test
rawdata = system.file('extdata/instruments/ecis_TimeResample.abp', package = 'vascr')
modeleddata = system.file('extdata/instruments/ecis_TimeResample_RbA.csv', package = 'vascr')
sampledefine = system.file('extdata/instruments/eciskey.csv', package = 'vascr')

#Then run the import
data1 = vascr_import("ECIS", model = modeleddata, key = sampledefine, experimentname = "TEST1")
data2 = vascr_import("ECIS", raw = rawdata, key = sampledefine, experimentname = "TEST2")
data3 = vascr_import("ECIS",model = modeleddata, raw = rawdata, key = sampledefine, experimentname = "TEST3")

# cellZscope test
rawdata = system.file('extdata/instruments/zscoperaw.txt', package = 'vascr')
modeleddata = system.file('extdata/instruments/zscopemodel.txt', package = 'vascr')
sampledefine = system.file('extdata/instruments/zscopekey.csv', package = 'vascr')

data4 = vascr_import("cellZscope", model = modeleddata, key = sampledefine, experimentname = "TEST4")
data5 = vascr_import("cellZscope", raw = rawdata, key = sampledefine, experimentname = "TEST5")
data6 = vascr_import("cellZscope",model = modeleddata, raw = rawdata, key = sampledefine, experimentname = "TEST6")

# xCELLigence test
rawdata = system.file('extdata/instruments/xcell.plt', package = 'vascr')
sampledefine = system.file('extdata/instruments/xcellkey.csv', package = 'vascr')

}
