% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_instrument_list}
\alias{vascr_instrument_list}
\title{Returns a list of the instruments supported by the VASCR package}
\usage{
vascr_instrument_list()
}
\value{
A vector of all the supported instrument names
}
\description{
Returns a list of the instruments supported by the VASCR package
}
\examples{
# vascr_instrument_list()
}
\keyword{internal}
