% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_instrument_units}
\alias{vascr_instrument_units}
\title{Return the units created by a certain instrument}
\usage{
vascr_instrument_units(instrument)
}
\arguments{
\item{instrument}{The instrument to find the units for}
}
\value{
a vector of units provided by an instrument
}
\description{
Return the units created by a certain instrument
}
\examples{
vascr_instrument_units("ECIS")
vascr_instrument_units("xCELLigence")
vascr_instrument_units("cellZscope")

}
