% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_levene}
\alias{vascr_levene}
\title{Run a Levene's test of normailty on a vascr dataset}
\usage{
vascr_levene(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{vascr dataset to analyse}

\item{unit}{unit to plot}

\item{frequency}{frequency to plot}

\item{time}{time to plot}

\item{priority}{vascr priority list}
}
\value{
A Levene Test object
}
\description{
Runs a test for homogenicity of variance, to ensure that the conditions of an ANOVA are met. Built into anova plotting functions and anova summary.
}
\examples{
# vascr_levene(growth.df, "R", 4000, 100)

}
