% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_lm}
\alias{vascr_lm}
\title{Generate a linear model of vascr data}
\usage{
vascr_lm(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{the dataset to use}

\item{unit}{Unit to select}

\item{frequency}{Frequency to select}

\item{time}{Time to select}

\item{priority}{override the default vascr priority list}
}
\value{
A linear model object
}
\description{
Generate a linear model of vascr data
}
\examples{
# vascr_lm(growth.df, "R", 4000, 100)

}
\keyword{internal}
