% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_save.R,
%   R/vascr_save.sync-conflict-20260526-135749-YEPQG6Q.R
\name{vascr_load}
\alias{vascr_load}
\title{Load a vascr dataset}
\usage{
vascr_load(path)

vascr_load(path)
}
\arguments{
\item{path}{the path to a .vascr file containing the saved dataset}
}
\value{
A vascr dataset

A vascr dataset
}
\description{
Load a vascr dataset

Load a vascr dataset
}
\examples{

path = system.file("extdata/test.vascr", package = "vascr")
vascr_load(path)


path = system.file("extdata/test.vascr", package = "vascr")
vascr_load(path)

}
