% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_make_name}
\alias{vascr_make_name}
\title{Title}
\usage{
vascr_make_name(
  data.df,
  select_cols = NULL,
  remove_blank = TRUE,
  fill_blank = "Control"
)
}
\arguments{
\item{data.df}{}

\item{select_cols}{}

\item{remove_blank}{}

\item{fill_blank}{}
}
\value{

}
\description{
Title
}
