% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{vascr_make_numeric}
\alias{vascr_make_numeric}
\title{Make a vector containing numbers with commas numeric}
\usage{
vascr_make_numeric(vector)
}
\arguments{
\item{vector}{The vector to process}
}
\value{
A numeric vector
}
\description{
Make a vector containing numbers with commas numeric
}
