% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_make_panel}
\alias{vascr_make_panel}
\title{Combine multiple VASCR plots into a single panel}
\usage{
vascr_make_panel(..., plots = NULL, legend_from_index = 1, ncols = 1)
}
\arguments{
\item{...}{The plots that need to be combined}

\item{plots}{A vector of plots, if not available in ... format}

\item{legend_from_index}{Which plot in the list to clone the legend from}

\item{ncols}{Number of colums}
}
\value{
A panel with multiple plots on it
}
\description{
Combine multiple VASCR plots into a single panel
}
\examples{
#plots = list()

}
