% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_anova.R
\name{vascr_make_significance_table}
\alias{vascr_make_significance_table}
\title{Make a dataframe of what is statistically significant}
\usage{
vascr_make_significance_table(
  data.df,
  time,
  unit,
  frequency,
  priority,
  confidence = 0.95,
  format = "toplot"
)
}
\arguments{
\item{data.df}{The dataframe to analyse}

\item{time}{The time to analyse}

\item{unit}{The unit to analyse}

\item{frequency}{The frequency to analyse}

\item{confidence}{The confidence level to analyse - default is 0.95}

\item{format}{The format to return the data frame in}
}
\value{
A table of what is significant
}
\description{
Make a dataframe of what is statistically significant
}
\examples{
#vascr_make_significance_table(growth.df, 50, "R", 4000, 0.95)
#vascr_make_significance_table(growth.df, 50, "R", 4000, 0.95, format = "Tukey_data")

}
