% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_match}
\alias{vascr_match}
\title{Match a string with the closest available option}
\usage{
vascr_match(match, vector)
}
\arguments{
\item{vector}{The vector to match into}

\item{tomatch}{The string to match}
}
\value{
A character string of the closest matched string
}
\description{
Match a string with the closest available option
}
\examples{
# vector = vascr_find_unit(growth.df, "all")
# vascr_match("Re", vector)
# vascr_match("Rb", vector)
# vascr_match(c("Rb", "Cm"), vector)
}
\keyword{internal}
