% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_median_well}
\alias{vascr_median_well}
\title{Calculate the median well in a set of wells}
\usage{
vascr_median_well(wells)
}
\arguments{
\item{wells}{A vector of wells to find the median of}
}
\value{
The name of the median well
}
\description{
This function finds the well that is the median of a set. This will be the most spacially central well on a plate. Using median eliminates the risk of well locations clashing, as the returned well will always be one of the set input. This also eliminates the noise associated with single replicates that need to be moved to the edge of a plate for technical reasons, however it will also mask that this movement has happened.
}
\details{
Works for both vertical, horrosontal and diffuse well configurations
}
\examples{
#vascr_median_well (c("A1", "B2", "C3"))
#vascr_median_well(c("A1", "NA", "NA", "NA"))
}
