% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tests.R
\name{vascr_missing_wells}
\alias{vascr_missing_wells}
\title{Check if there are any standard 96 well plate wells missing from a 96 well plate}
\usage{
vascr_missing_wells(wells)
}
\arguments{
\item{wells}{A vector of wells to test, or a vascr dataset that will be truncated}
}
\value{
A vector of the well names that are missing from the dataset
}
\description{
Check if there are any standard 96 well plate wells missing from a 96 well plate
}
\keyword{internal}
