% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_subset.R
\name{vascr_modeled_in_data}
\alias{vascr_modeled_in_data}
\title{Find which modeled units in the dataset}
\usage{
vascr_modeled_in_data(data)
}
\arguments{
\item{data}{The dataset to analyse}
}
\value{
A vector of the modeled units in the dataset
}
\description{
Find which modeled units in the dataset
}
\examples{
#vascr_modeled_in_data(growth.df)


}
