% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_multiplot}
\alias{vascr_multiplot}
\title{Title}
\usage{
vascr_multiplot(data, unit, frequency = 0, time)
}
\arguments{
\item{data}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_multiplot(data = growth.df, unit = "raw", time = 100)

}
