% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_odd_in_vector}
\alias{vascr_odd_in_vector}
\title{Checks if the value is placed in an odd position in a vector}
\usage{
vascr_odd_in_vector(value, vector)
}
\arguments{
\item{value}{The value to search for}

\item{vector}{Vector to search}
}
\value{
A boolean
}
\description{
Checks if the value is placed in an odd position in a vector
}
\examples{
# vector = vascr_units_table()$Unit
# vascr_odd_in_vector("Alpha", vector)
# vascr_odd_in_vector("Cm", vector)
# vascr_odd_in_vector("Drift", vector)


}
\keyword{internal}
