% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_plot}
\alias{vascr_plot}
\title{ECIS plot}
\usage{
vascr_plot(
  data,
  unit = "R",
  frequency = 4000,
  replication = "summary",
  time = Inf,
  samplecontains = "",
  experiment = "",
  error = Inf,
  linesize = 1,
  normtime = NULL,
  divide = FALSE,
  errorsize = 1,
  alphavalue = 0.1,
  confidence = 1,
  xlab = "Time (hours)",
  ylab = "Value",
  title = "Title",
  stripidentical = TRUE,
  cols = NULL,
  verbose = TRUE,
  preprocessed = FALSE,
  continuous = NULL,
  alignkey = NULL,
  continuouscontains = NULL,
  returndata = FALSE,
  sortkeyincreasing = TRUE,
  showpoints = FALSE
)
}
\arguments{
\item{data}{A standard ECIS data frame to plot}

\item{unit}{Unit to plot}

\item{frequency}{The frequency of data to display. All modelled variables have a frequency of 0}

\item{replication}{How much of the replicaiton to display. Options are 'all', 'experiment', 'summary" or "plate"}

\item{time}{The time to subset if a slice is required. If set to Inf all data will be displayed}

\item{samplecontains}{Optional, only samples that contain this string will be plotted. Standard search and wildcard}

\item{experiment}{Optional, allows you to limit the experiment plotted. Experiment names should be separated with |
searches apply.}

\item{error}{Display error bars. 0 displays no error bars, 1 displays them all. Higher numbers will reduce the frequency of error bars plotted.}

\item{linesize}{Width of mean lines shown on graphs}

\item{normtime}{The time to normalise the data to}

\item{divide}{Selector for if normalisation should be done by division or subtraction. Default is subtraction (FALSE)}

\item{errorsize}{Width of error bars shown on graphs}

\item{alphavalue}{Alpha value of area enclosed by error bars. May be lowered for buisy graphs}

\item{confidence}{The confidence to use when generating basic significance plots}

\item{xlab}{X axis value}

\item{ylab}{Y axis value}

\item{title}{The title of the ggplot}

\item{stripidentical}{Remove cols from the data where all data points are identical. Default is true}

\item{cols}{The columns of data to display in the names of plots. Allows for shortening of names where the automatic stripidentical fails}

\item{verbose}{Outputs where in the funtion ploting is happening. Usefull for debugging as needed}

\item{preprocessed}{Selects if the data is preprocessed. Will not run explosion and implosion if that is the case}

\item{continuous}{Selects a continuous point for plotting}

\item{alignkey}{Aligns key points (max, min ect). See vascr_align_key for details}

\item{continuouscontains}{Subset of samplecontains, only returns data where the continuous value contains this data}

\item{returndata}{Return the dataset, rather than the graph. Default FALSE, usefull for debugging}

\item{showpoints}{Show the time points on the graph}
}
\value{
A ggplot2 object
}
\description{
The master function for generating ggplot2 objects from ECIS data. Incorporates all the summarising, testing and graphing operations into a single function.
}
\examples{
growth.df$Instrument = "ECIS"

}
