% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R, R/vascr_plot_anova.R
\name{vascr_plot_anova}
\alias{vascr_plot_anova}
\title{Make a display with all the anova analyasis pre-conducted}
\usage{
vascr_plot_anova(
  data.df,
  unit,
  frequency,
  time,
  visualisation = NULL,
  priority = NULL,
  ...
)

vascr_plot_anova(
  data.df,
  unit,
  frequency,
  time,
  visualisation = NULL,
  priority = NULL,
  ...
)
}
\arguments{
\item{data.df}{A standard ECIS dataset.}

\item{unit}{Unit to be analysed.}

\item{frequency}{Frequency to be used. All modeled units have a frequency of 0.}

\item{time}{Time to be used.}

\item{priority}{analyasis priority. Will use default if not filled in.}

\item{...}{other controls to be pushed to publish plot and generate graphdata}

\item{data}{vascr dataset to plot}
}
\value{
A matrix of different ANOVA tests

ANOVA results printed to the console.
}
\description{
This function takes a standard ECIS dataset and runs two way ANOVA on it based on the variaiton between treatments and experiments. It then runs a post-hoc test to determine which pairs of groups are statistically different. Results are printed to the console as text.
}
\details{
This analyasis requires multiple experiments, as working out the differences in a single experiment is not a good idea because it does not account for experimental variaiton, making the analyasis weaker and more unreliable as results may not be repeatable.

Firstly, a histogram is generated for a visual check of the data. This should mimic what you are expecting.

To interpret the data first check that the data is roughly linear on the Shapiro-Wilk normailty graph,and that the p<0.05. This is more important for smaller datasets, and becomes irrelevant once n>9. Therefore you can be relativley non-stringent when applying this test to three independent experiments, where each experiment had triplicates. Also check that the residuals on the right hand graph are roughly normal, this is easier for larger datasets.
}
\examples{

vascr_plot_anova(growth.df, "R", 4000, 100)

vascr_plot_anova(growth.df, unit = 'R', frequency = 4000,time = 50)


}
