% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_anova_bar_reference}
\alias{vascr_plot_anova_bar_reference}
\title{Title}
\usage{
vascr_plot_anova_bar_reference(
  data.df,
  unit,
  frequency,
  time,
  reference = min(data.df$SampleID),
  breaklines = TRUE
)
}
\arguments{
\item{data.df}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}

\item{reference}{}
}
\description{
Title
}
