% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_plot_badwell_plate}
\alias{vascr_plot_badwell_plate}
\title{Title}
\usage{
vascr_plot_badwell_plate(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_plot_badwell_plate(growth.df)

}
