% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_bar}
\alias{vascr_plot_bar}
\title{Plot a vascr data set as a bar}
\usage{
vascr_plot_bar(data, priority = NULL, error = Inf, ...)
}
\arguments{
\item{data}{}

\item{...}{Any argument to be passed to vascr_polish_plot or vascr_prep_graphdata}
}
\value{

}
\description{
Plot a vascr data set as a bar
}
\examples{

vascr_plot_bar(growth.df, level = "experiments", frequency = 4000, unit = "R", time = list(50,100), error = Inf)

vascr_plot_bar(growth.df, level = "wells", frequency = 4000, unit = "R", time = 50, error = Inf)
vascr_plot_bar(growth.df, level = "experiments", frequency = 4000, unit = "R", time = 50, error = Inf)
vascr_plot_bar(growth.df, level = "summary", frequency = 4000, unit = "R", time = 50, error = Inf)
vascr_plot_bar(growth.df, level = "deviation", frequency = 4000, unit = "R")

}
