% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova.R
\name{vascr_plot_bar_dunnett}
\alias{vascr_plot_bar_dunnett}
\title{Title}
\usage{
vascr_plot_bar_dunnett(data.df, unit, frequency, time, reference, stars = TRUE)
}
\arguments{
\item{data.df}{a vascr dataframe}

\item{unit}{the unit to use}

\item{frequency}{the frequency to use}

\item{time}{the time to use for the bar plot and ANOVA}

\item{reference}{SampleID of the sample to use as the reference for statistical analysis}

\item{stars}{Show stars, or rounded P values}
}
\description{
Title
}
\examples{
vascr_plot_bar_dunnett(growth.df, "R", 4000, 50, "0_cells + hCMEC/d3_line")

}
