% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_plot_blank}
\alias{vascr_plot_blank}
\title{Return a blank white box}
\usage{
vascr_plot_blank()
}
\value{
A blank ggplot2 object
}
\description{
Return a blank white box
}
\examples{
# vascr_plot_blank()
}
\keyword{internal}
