% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_box_replicate}
\alias{vascr_plot_box_replicate}
\title{Plot replicate data sets as a box plot}
\usage{
vascr_plot_box_replicate(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{The dataset to plot}

\item{unit}{Unit to plot}

\item{frequency}{Frequency to plot}

\item{time}{Time to plot}

\item{priority}{vascr priority, if empty default will be used}
}
\value{
A ggplot2 box plot of replicate experiments
}
\description{
Plot replicate data sets as a box plot
}
\examples{
# vascr_plot_box_replicate(growth.df, "R", 4000, 100)

}
\keyword{internal}
