% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_cross_correlation.R
\name{vascr_plot_cc}
\alias{vascr_plot_cc}
\title{Title}
\usage{
vascr_plot_cc(data.df)
}
\description{
Title
}
\examples{
data.df = vascr::growth.df \%>\% 
            vascr_subset(unit = "R", frequency = 4000, sampleid = c(1,4,7,8), time = c(1,20)) \%>\%
            vascr_cc()

# data.df \%>\% vascr_plot_cc()
# data.df \%>\% vascr_summarise_cc("experiments") \%>\% vascr_plot_cc()
# data.df \%>\% vascr_summarise_cc("summary") \%>\% vascr_plot_cc()

}
