% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_cross_correlation.R
\name{vascr_plot_cc_stats}
\alias{vascr_plot_cc_stats}
\title{Title}
\usage{
vascr_plot_cc_stats(data.df, unit = "R", frequency = 4000)
}
\arguments{
\item{data.df}{}
}
\description{
Title
}
\examples{

vascr_plot_cc_stats(growth.df)

}
