% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_column}
\alias{vascr_plot_column}
\title{Title}
\usage{
vascr_plot_column(data, priority = NULL, ...)
}
\arguments{
\item{data}{}

\item{replication}{}

\item{error}{}

\item{title}{}

\item{xlab}{}

\item{ylab}{}

\item{time}{}

\item{unit}{}

\item{frequency}{}

\item{confidence}{}
}
\value{

}
\description{
Title
}
\examples{

data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "summary", time = list(50))
vascr_plot_column(data)

data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "experiments", time = list(50,100))
vascr_plot_column(data)

data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "wells", samplecontains = c("35,000", "30,000"), time = list(50,100,150))
vascr_plot_column(data, priority = c("Experiment","Time", "Sample", "Experiment"))

unique(data$Well) 

}
