% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{vascr_plot_continuous}
\alias{vascr_plot_continuous}
\title{Plot a categorical ECIS variable}
\usage{
vascr_plot_continuous(data, cols, continuous, ...)
}
\arguments{
\item{data}{A standard ECIS dataset}

\item{cols}{The column names to display}

\item{continuous}{The nominated continuous variable}

\item{unit}{The unit to plot}

\item{frequency}{The frequency to plot}

\item{replication}{The replication level to plot at}

\item{time}{The time to run this analyasis at}

\item{error}{How much error to display}

\item{alphavalue}{The alpha of the error bars (how clear they are so you can see the overlap)}

\item{xlab}{Label on the X axis}

\item{ylab}{Label on the Y axis}

\item{title}{The title}
}
\value{
a ggplot 2 object
}
\description{
Plot a categorical ECIS variable
}
\examples{

growth.df$Instrument = "ECIS"

vascr_plot_continuous(growth.df, "R", 4000, "wells", 50, Inf, continuous = "cells")
# vascr_plot_continuous(growth.df, "R", 4000, "experiments", 50, continuous = "cells")
# vascr_plot_continuous(growth.df, "R", 4000, "summary", 50, continuous = "cells", error = Inf)


}
