% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_quality.R
\name{vascr_plot_deviation}
\alias{vascr_plot_deviation}
\title{Various plots that show the deviation between the median well and various datapoints}
\usage{
vascr_plot_deviation(
  data,
  max_deviation = 0,
  deviation = 0,
  priority = NULL,
  unit = "R",
  frequency = 4000,
  visualisation = NULL,
  ...
)
}
\arguments{
\item{data}{A vascr dataset to show}

\item{max_deviation}{The maximum deviation of interest}

\item{deviation}{The maximum allowable deviation at any one time point. Defaults to the same as max_deviation.}

\item{priority}{The priority of vascr (or other) variables should be displayed in. See vascr_priority for details.}

\item{unit}{The unit you wish to run the analyasis on. A raw unit is usually best. Default is R.}

\item{frequency}{The frequency you want to run the analyasis on. Default is 4000.}

\item{visualisation}{The visualisation you want to focus in on. Options are "bar", "plate" or "line". The default is to tile all three.}

\item{...}{Any other arguements to be passed on to either vascr_subset or vascr_polish_plot}
}
\value{
A ggplot object, or matrix of ggplot objects
}
\description{
This function by default outputs three sets of diagnostic plots that describe where, by how much and when a well has drifted from the median for that dataset. Individual visualisations from this set can also be requested using the visualisation paramater. The max_deviation and deviation paramaters can be used to annotate these plots, allowing acceptable variation.
}
\examples{
vascr_plot_deviation(growth.df)
vascr_plot_deviation(growth.df, visualisation = "bar")
vascr_plot_deviation(growth.df, visualisation = "plate")
vascr_plot_deviation(growth.df, visualisation = "line")
vascr_plot_deviation(growth.df, max_deviation = 0.2)

}
