% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_plot_frequency.R
\name{vascr_plot_frequency}
\alias{vascr_plot_frequency}
\title{Generate frequency plots}
\usage{
vascr_plot_frequency(data, priority, ...)
}
\arguments{
\item{data}{A vascr formatted datapoint}

\item{unit}{The unit to plot}

\item{time}{The time to plot at}

\item{frequency}{The frequency to plot}

\item{samplecontains}{Subset by sample containing frequency}

\item{experiment}{}

\item{error}{}

\item{alignkey}{}

\item{normtime}{}

\item{preprocessed}{}

\item{continuouscontains}{}

\item{stripidentical}{}
}
\value{

}
\description{
Library of pre-formed ggplot functions that generate frequency vs value plots for a range of different combinations of instruments, samples and experimental replication levels. Indexed with conditioned data from vascr_plot.
}
\examples{

vascr_plot_frequency(growth.df, replication = "summary", unit = "C", errorbars = Inf, samplecontains = "35,000")
vascr_plot_frequency(growth.df, replication = "experiments", unit = "C", errorbars = 1)
vascr_plot_frequency(growth.df, replication = "wells", unit = "C", errorbars = 1)
vascr_plot(growth.df)

vascr_plot_frequency(time = 100)

data = growth.df
replication = "wells"

}
