% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_graphics.R,
%   R/vascr_graphics.sync-conflict-20260526-135829-YEPQG6Q.R
\name{vascr_plot_grid}
\alias{vascr_plot_grid}
\title{Plot out each replicate well in a grid, with QC overlays}
\usage{
vascr_plot_grid(data.df, threshold = 0.2)

vascr_plot_grid(data.df, threshold = 0.2)
}
\arguments{
\item{data.df}{a vascr formatted data frame of single values}

\item{threshold}{threshold at which a data point is determined to be an outlier}
}
\value{
A plot to be used for QC

A plot to be used for QC
}
\description{
Plot out each replicate well in a grid, with QC overlays

Plot out each replicate well in a grid, with QC overlays
}
\examples{
grid.df = growth.df \%>\% vascr_subset(unit = "R", frequency = "4000", experiment  = 1)
vascr_plot_grid(grid.df)

grid.df = growth.df \%>\% vascr_subset(unit = "R", frequency = "4000", experiment  = 1)
vascr_plot_grid(grid.df)

}
