% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_heatmap}
\alias{vascr_plot_heatmap}
\title{Title}
\usage{
vascr_plot_heatmap(data, ...)
}
\arguments{
\item{data}{}

\item{time}{}

\item{unit}{}

\item{frequency}{}
}
\value{

}
\description{
Title
}
\examples{

vascr_plot_heatmap(growth.df, time = 100, unit = "Rb")
vascr_plot_heatmap(growth.df, 100, "R", 4000)

}
