% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_plot_isolate}
\alias{vascr_plot_isolate}
\title{Graph a single well that is misbehaving}
\usage{
vascr_plot_isolate(data.df, well, ...)
}
\arguments{
\item{data.df}{The dataset the well is in}

\item{well}{The well to be isolated}

\item{...}{Other conditions to pass on to the vascr_plot command that generates the graph}
}
\value{
A ggplot graph showing the isoated well and the experimental median well
}
\description{
Graph a single well that is misbehaving
}
\examples{

#datum = vascr_plot_isolate(growth.df, well = "A3", unit = "R", frequency = 4000,
# error = Inf, priority = NULL)

#vascr_plot_line(datum, preprocessed = TRUE, priority = c("Time", "Sample"))

}
\keyword{internal}
