% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_line.R
\name{vascr_plot_keylines}
\alias{vascr_plot_keylines}
\title{Title}
\usage{
vascr_plot_keylines(
  plot,
  key_events,
  labelx = TRUE,
  linetype = "dashed",
  linesize = 1
)
}
\arguments{
\item{plot}{}

\item{key_events}{}
}
\value{

}
\description{
Title
}
