% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_line.R
\name{vascr_plot_keyrange}
\alias{vascr_plot_keyrange}
\title{Title}
\usage{
vascr_plot_keyrange(plot, key_events, labelx = TRUE)
}
\arguments{
\item{plot}{}

\item{key_events}{}
}
\value{

}
\description{
Title
}
