% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_levene}
\alias{vascr_plot_levene}
\title{Plot the results of a Levene's test with a fitted variables and residuals plot}
\usage{
vascr_plot_levene(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{a vascr dataset}

\item{unit}{Unit to analyse}

\item{frequency}{Frequency to analyase}

\item{time}{Time to analyse}

\item{priority}{A vascr list of priorities. If left blank default will be used.}
}
\value{
A ggplot of a levene's test and the underlying data analysed
}
\description{
Plot the results of a Levene's test with a fitted variables and residuals plot
}
\examples{
vascr_plot_levene(growth.df, "R", 4000, 100)


}
