% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_line}
\alias{vascr_plot_line}
\title{Title}
\usage{
vascr_plot_line(data, priority = NULL, ...)
}
\arguments{
\item{data}{}

\item{replication}{}

\item{error}{}

\item{title}{}

\item{xlab}{}

\item{ylab}{}

\item{linesize}{}

\item{alphavalue}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_plot_line(growth.df, priority = "Sample", unit = "R", frequency = 4000, replication = "summary")


}
