% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_line.R
\name{vascr_plot_line_panel}
\alias{vascr_plot_line_panel}
\title{Title}
\usage{
vascr_plot_line_panel(data.df, key_points = NULL)
}
\arguments{
\item{data.df}{}

\item{key_points}{}
}
\description{
Title
}
