% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_plot_matrix}
\alias{vascr_plot_matrix}
\title{Plot the wells of a 96 well plate}
\usage{
vascr_plot_matrix(data.df)
}
\arguments{
\item{data.df}{A standard ECIS dataframe. Ideally this will contain only one experiment, but multiple experiments are supported.}

\item{unit}{The unit to plot. Default is R}

\item{frequency}{The frequency to plot. Default is 4000}

\item{...}{Values to be passed onto prep_graphdata and polish_plot}
}
\value{
A GGplot2 matrix
}
\description{
Plot the wells of a 96 well plate
}
\examples{
#vascr_plot_matrix(growth.df, unit = "Rb")

}
\keyword{internal}
