% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_plot_model}
\alias{vascr_plot_model}
\title{Plot all values generated by the ECIS model}
\usage{
vascr_plot_model(alldata.df, ...)
}
\arguments{
\item{alldata.df}{An ECIS data frame}

\item{...}{Any other arguements from vascr_plot that you might want passed through to all of the graphs}
}
\value{
A matrix containing graphs of all the variables generated by the ECIS model.
}
\description{
This is a pre-generated piece of code that automaticaly plots all the variables generated by the ECIS model against time. Resistance at 4000 Hz is also included as a sensible sanity check of the model.
}
\examples{

vascr_plot_model(growth.df, replication = "wells")

}
