% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_normality}
\alias{vascr_plot_normality}
\title{Plot rediduals overlaid with a normal curve}
\usage{
vascr_plot_normality(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{The dataset to plot}

\item{unit}{Unit to plot}

\item{frequency}{Frequency to plot}

\item{time}{Time to plot}

\item{priority}{Vascr priority list, will use the default if available}
}
\value{
a ggplot with rediduals overlaid by a normal curve
}
\description{
Plot rediduals overlaid with a normal curve
}
\examples{

vascr_plot_normality(growth.df, "R", 4000, 100)

}
