% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_qq}
\alias{vascr_plot_qq}
\title{Generate a qq plot and shapiro test from a vascr data frame}
\usage{
vascr_plot_qq(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{vascr dataset to use}

\item{unit}{Unit to return}

\item{frequency}{Frequency to return}

\item{time}{Timepoint to use}

\item{priority}{vascr priority list for analyasis, if blank default will be used}
}
\value{
A ggpubr ggqqplot object
}
\description{
Generate a qq plot and shapiro test from a vascr data frame
}
\examples{
# vascr_plot_qq(growth.df, "R", 4000, 100)

}
\keyword{internal}
