% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{vascr_plot_resample_range}
\alias{vascr_plot_resample_range}
\title{Title}
\usage{
vascr_plot_resample_range(data.df, unit = "R", frequency = 4000, well = "A01")
}
\arguments{
\item{data.df}{}
}
\description{
Title
}
\examples{
vascr_plot_resample_range(growth.df)

bigdata = vascr_import("ECIS", raw ="raw_data/growth/growth1_raw.abp", experiment = 1)

toprocess = bigdata \%>\% mutate(Experiment = "1", Sample = "Test")

vascr_plot_resample_range(data.df = toprocess)

}
