% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_samplemap}
\alias{vascr_plot_samplemap}
\title{ECIS plot samplemap}
\usage{
vascr_plot_samplemap(data, title = "Title", stripidentical = TRUE)
}
\arguments{
\item{data}{The datapoint to plot}

\item{title}{Title to be placed on the graph}
}
\value{

}
\description{
ECIS plot samplemap
}
\examples{

vascr_plot_samplemap(growth.df)

}
