% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_plate.R
\name{vascr_plot_structure}
\alias{vascr_plot_structure}
\title{ECIS plot samplemap}
\usage{
vascr_plot_structure(data, title = "Title", stripidentical = TRUE, ...)
}
\arguments{
\item{data}{The datapoint to plot}

\item{title}{Title to be placed on the graph}
}
\value{

}
\description{
ECIS plot samplemap
}
\examples{

vascr_plot_structure(growth.df)

}
