% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_time_vline}
\alias{vascr_plot_time_vline}
\title{Plot a line graph with a vertical line on it}
\usage{
vascr_plot_time_vline(data.df, unit, frequency, time, priority = NULL, ...)
}
\arguments{
\item{data.df}{A vascr dataset}

\item{unit}{Unit to plot}

\item{frequency}{Frequency to plot}

\item{time}{Time to plot}

\item{priority}{Vascr priority list. Blank will use the baked in default.}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Plot a line graph with a vertical line on it
}
\examples{
vascr_plot_time_vline(growth.df, "R", 4000, 100)

}
