% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_polish_plot}
\alias{vascr_polish_plot}
\title{Post-processing on ECIS plots to make them a bit more pretty}
\usage{
vascr_polish_plot(plot, rotate_x_angle = 45, logscale = "", title = NULL)
}
\arguments{
\item{plot}{The plot to format}

\item{rotate_x_angle}{Angle by which X axis values should be rotated}

\item{logscale}{Should X, Y or XY axes be log10 transformed}

\item{title}{Title of the graph}
}
\value{
A standardised ggplot2 object
}
\description{
Post-processing on ECIS plots to make them a bit more pretty
}
\examples{
# Automatically applied to ggplot, should not be required externaly

#plot = vascr_plot(growth.df)
#vascr_polish_plot(plot)
#vascr_polish_plot(plot, rotate_x_angle = FALSE)

}
\keyword{internal}
