% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_prep_graphdata}
\alias{vascr_prep_graphdata}
\title{Prepare a dataset to be graphed by vascar_graph_xxx}
\usage{
vascr_prep_graphdata(
  data.df,
  unit = "",
  frequency = Inf,
  time = NULL,
  samplecontains = NULL,
  experiment = NULL,
  error = Inf,
  alignkey = NULL,
  normtime = NULL,
  divide = FALSE,
  preprocessed = FALSE,
  continuouscontains = NULL,
  stripidentical = TRUE,
  sortkeyincreasing = TRUE,
  level = "summary",
  errortype = "sem",
  subsample = NULL
)
}
\arguments{
\item{unit}{Unit to subset to}

\item{frequency}{Frequency to subset to}

\item{time}{Time to subset to}

\item{samplecontains}{Subset only sample names that contain this string}

\item{experiment}{Experiment to subset to}

\item{error}{How much error to plot. Required to allow subsampling if required}

\item{alignkey}{Should key points be aligned}

\item{normtime}{Time to normalise to}

\item{divide}{Should normalisation be by division (true) or subtraction (false)}

\item{preprocessed}{Is the data already processed and therefore should be left alone}

\item{continuouscontains}{Subset variables where the sample contains this string}

\item{stripidentical}{Should entireley identical columns be removed}

\item{sortkeyincreasing}{Should samples be sorted in an increasing way}

\item{level}{The level of summary to return}

\item{errortype}{SEM or SD errors to generate}

\item{data}{Vascr dataset to plot}
}
\value{
A vascr dataset prepared for use in graphing
}
\description{
Central data subset, cleanup and label prep function for generation of graphics
}
\examples{

vascr_plot(growth.df, unit = "Rb", level = "experiments", frequency = 0)

datum2 = vascr_prep_graphdata(growth.df, unit = "Rb", level = "summary", frequency = 0)


#data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "experiments")
#data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "wells")


}
