% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_prep_statdata}
\alias{vascr_prep_statdata}
\title{Prep data for statistical analyasis}
\usage{
vascr_prep_statdata(data.df, unit, frequency, time)
}
\arguments{
\item{data.df}{The dataset to analyse}

\item{unit}{The unit to analyse}

\item{frequency}{Frequency to use}

\item{time}{Timepoint to use}
}
\value{
A conditioned vascr dataset
}
\description{
Prep data for statistical analyasis
}
\examples{
#vascr_prep_statdata(growth.df, "R", 4000, 100)

}
\keyword{internal}
