% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_subset.R
\name{vascr_raw_in_data}
\alias{vascr_raw_in_data}
\title{Return the raw units in the dataset}
\usage{
vascr_raw_in_data(data)
}
\arguments{
\item{data}{The dataset to search}
}
\value{
A vector containing the raw units present in the data
}
\description{
Return the raw units in the dataset
}
\examples{
# vascr_raw_in_data(growth.df)

}
