% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_realise_frequencies}
\alias{vascr_realise_frequencies}
\title{Title}
\usage{
vascr_realise_frequencies(data, frequency)
}
\arguments{
\item{data}{}

\item{frequency}{}
}
\value{

}
\description{
Title
}
