% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_realise_units}
\alias{vascr_realise_units}
\title{DEPRICATED: alias for vascr_find unit}
\usage{
vascr_realise_units(...)
}
\arguments{
\item{...}{variables for vascr_find_unit}
}
\value{
results from vascr_find_unit
}
\description{
DEPRICATED: alias for vascr_find unit
}
\examples{
# See vascr_find_unit
}
\keyword{internal}
