% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_remove_from_vector}
\alias{vascr_remove_from_vector}
\title{Title}
\usage{
vascr_remove_from_vector(values, vector)
}
\arguments{
\item{values}{}

\item{vector}{}
}
\value{

}
\description{
Title
}
