% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_remove_metadata}
\alias{vascr_remove_metadata}
\title{Remove all non-core ECIS data from a data frame}
\usage{
vascr_remove_metadata(data.df, subset = "all")
}
\arguments{
\item{data.df}{An ECIS dataset}

\item{subset}{What to strip off. Default is all, more options to come.}
}
\value{
A dataset containing only the core ECIS columns
}
\description{
Usefull if you want to to further data manipluation, without having to worry about tracking multiple, unknown columns.
}
\examples{
#growth.df$Instrument = "ECIS"
#exploded.df = vascr_explode(growth.df)
#cleaned.df = vascr_remove_metadata(exploded.df)
#identical(growth.df,cleaned.df)
}
\keyword{internal}
