% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_remove_stats}
\alias{vascr_remove_stats}
\title{Clean the statistics off a vascr dataset}
\usage{
vascr_remove_stats(data.df)
}
\arguments{
\item{data.df}{The dataset to clean}
}
\value{

}
\description{
Clean the statistics off a vascr dataset
}
\examples{
vascr_remove_stats(vascr_summarise(growth.df,level = "summary"))

}
