% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_rename_columns}
\alias{vascr_rename_columns}
\title{Title}
\usage{
vascr_rename_columns(data.df, old, new)
}
\arguments{
\item{data.df}{}

\item{old}{}

\item{new}{}
}
\value{

}
\description{
Title
}
