% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_resample_time}
\alias{vascr_resample_time}
\title{Title}
\usage{
vascr_resample_time(data.df, npoints = vascr_time_samples(data.df))
}
\arguments{
\item{data.df}{The vascr dataset to resample}

\item{npoints}{Manually specifiiy the number of points to resample at, default is the same frequency as the input dataset}
}
\description{
Title
}
