% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_sample_table}
\alias{vascr_sample_table}
\title{Title}
\usage{
vascr_sample_table(data.df)
}
\arguments{
\item{data.df}{}
}
\value{

}
\description{
Title
}
\examples{
data.df = xcell
samplecols = vascr_sample_table(data.df)
}
