% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_save.R,
%   R/vascr_save.sync-conflict-20260526-135749-YEPQG6Q.R
\name{vascr_save}
\alias{vascr_save}
\title{Save a vascr dataset}
\usage{
vascr_save(data.df, path)

vascr_save(data.df, path)
}
\arguments{
\item{data.df}{The vascr dataset to save}

\item{path}{The path to save the file to}
}
\value{
A .vascr file containing a vascr dataset

A .vascr file containing a vascr dataset
}
\description{
Save a vascr dataset

Save a vascr dataset
}
\examples{
path = tempfile()
vascr_save(growth.df, path = path)

path = tempfile()
vascr_save(growth.df, path = path)

}
