% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_shapiro}
\alias{vascr_shapiro}
\title{Plot a shapiro test}
\usage{
vascr_shapiro(data.df, unit, frequency, time, priority = NULL)
}
\arguments{
\item{data.df}{vascr dataset to analyse}

\item{unit}{Unit to plot}

\item{frequency}{Frequency to plot}

\item{time}{Time to plot}

\item{priority}{Vascr priority list}
}
\value{
A shapiro test of the selected data
}
\description{
Plot a shapiro test
}
\examples{
# vascr_shapiro(growth.df, "R", 4000, 100)

}
\keyword{internal}
