% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_shiny2.R
\name{vascr_shiny}
\alias{vascr_shiny}
\title{Launch the vascr UI}
\usage{
vascr_shiny(data.df)
}
\arguments{
\item{data.df}{Data to preload into shiny app}
}
\value{
A shinyApp to work with vascr data
}
\description{
Launch the vascr UI
}
\examples{
 if(interactive()){
   vascr_shiny()
 }
}
