% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_shorten_name}
\alias{vascr_shorten_name}
\title{Shorten a name down to only some variables, and write out those that change}
\usage{
vascr_shorten_name(
  name_vector,
  select_cols = NULL,
  remove_blank = TRUE,
  fill_blank = "Control",
  include_wells = FALSE
)
}
\arguments{
\item{name_vector}{A vector of names to implode}

\item{select_cols}{The columns to incorportate}

\item{remove_blank}{Should blank columns be removed, default TRUE}

\item{fill_blank}{Text to fill blanks in with, default "Control"}

\item{include_wells}{Should well id's be included}
}
\value{

}
\description{
Shorten a name down to only some variables, and write out those that change
}
