% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_strip_unit}
\alias{vascr_strip_unit}
\title{Strip units from a title containing string}
\usage{
vascr_strip_unit(string)
}
\arguments{
\item{string}{The string to strip from}
}
\value{
The trimmed string
}
\description{
Strip units from a title containing string
}
\examples{
string = c("nm.ml.Oleandrin", "nm.ATP")
vascr_strip_unit(string)
}
