% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_subset}
\alias{vascr_subset}
\title{Title}
\usage{
vascr_subset(
  data.df,
  time = NULL,
  unit = NULL,
  well = NULL,
  value_greater_than = NULL,
  value_less_than = NULL,
  value_equal_to = NULL,
  sample_contains = NULL,
  sample_not_contains = NULL,
  sample_equal_to = NULL,
  variable_set = NULL,
  variable_unset = NULL,
  variable_equal_to = NULL,
  variable_not_equal_to = NULL,
  variable_greater_than = NULL,
  variable_less_than = NULL,
  include_vehicle = TRUE,
  frequency = NULL,
  experiment = NULL,
  instrument = NULL,
  max_deviation = NULL,
  subsample = NULL,
  return_lists = FALSE
)
}
\arguments{
\item{data.df}{}

\item{time}{}

\item{unit}{}

\item{well}{}

\item{value_greater_than}{}

\item{value_less_than}{}

\item{value_equal_to}{}

\item{sample_contains}{}

\item{sample_not_contains}{}

\item{sample_equal_to}{}

\item{variable_set}{}

\item{variable_unset}{}

\item{variable_equal_to}{}

\item{variable_not_equal_to}{}

\item{variable_greater_than}{}

\item{variable_less_than}{}

\item{include_vehicle}{}

\item{frequency}{}

\item{experiment}{}

\item{instrument}{}

\item{max_deviation}{}

\item{subsample}{}

\item{return_lists}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_subset(growth.df)
vascr_subset(growth.df, time = 40)
vascr_subset(growth.df, time = NULL)
 
vascr_subset(growth.df, unit = "Rb")
vascr_subset(growth.df, unit = "R")
vascr_subset(growth.df, well = "A1")
vascr_subset(growth.df, value_less_than = 100)
vascr_subset(growth.df, max_deviation = 10)

vascr_subset(growth.df, time = c(5,20))

vascr_subset(czs)

}
