% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_subset_continuous}
\alias{vascr_subset_continuous}
\title{#' Resample ECIS data onto a constant time base
#' 
#' This will currently over-sample data, use with care
#'
#' @param data.df A standard ECIS data frame
#' @param by  The frequency at which to resample
#' @param from The value at which to start resampling
#' @param to  The max value to resample
#' @param zero_time The value that will be set as 0 after the normalisation is done. Usefull for aligning treatment times of multiple experiments.
#' 
#' @importFrom stringr str_remove
#' @importFrom tidyr unite spread gather separate
#' @importFrom stats approxfun
#' @importFrom magrittr "%>%"
#'
#' @return An ECIS dataset with re-located time points
#' 
#' @export
#'
#' @examples
#' 
#' #data = vascr_resample(data.df = growth.df, by = 10)
#' #head (data)
#' 
vascr_resample = function (data.df, by, from = -Inf, to = Inf, zero_time = 0)
{
  
  if(from == -Inf)
  {
    from = min(data.df$Time)
  }
  if(to == Inf)
  {
    to = max(data.df$Time)
  }
  
  if(from<min(data.df$Time))
  {
    warning(paste("From is below than the minimum time in the dataset. Please select a number above", min(data.df$Time)))
  }
  
  if(to>max(data.df$Time))
  {
    warning(paste("To is greater than the maximum of the dataset. Please select a number below", max(data.df$Time)))
  }
  
  movedata = vascr_remove_metadata(data.df)
  movedata = vascr_subset(movedata, time = c(from,to))
  
  movedata$Time = movedata$Time - zero_time
  
  combinedcolnames = colnames(movedata)
  cleancolnames = str_remove(combinedcolnames, "Time")
  cleancolnames = str_remove(cleancolnames, "Value")
  cleancolnames = cleancolnames[cleancolnames!= ""]
  
  movedata = unite(movedata, col = "Stream", -Value, -Time, sep = "#")
  movedata = unique(movedata)
  movedata2 = spread(movedata, Stream, Value)
  
  # Generate new times
  newtimepoints= seq(from=from,to=to,by=by)
  oldtimepoints = movedata2$Time
  
  # Construct an empty data frame to take the new data
  movedata3 <- movedata2[0,]
  movedata3 = as.data.frame(movedata3)
  movedata3[nrow(movedata3)+length(newtimepoints),] <- NA
  movedata3$Time = newtimepoints
  
  currentcol = 2
  totalcols = length(colnames(movedata2))
  
  movedata2 = as.data.frame(movedata2)
  
  while(currentcol<(totalcols+1))
  {
    oldtp = as.vector(oldtimepoints)
    oldval = as.vector(movedata2[,currentcol])
    replot = approxfun(oldtp, oldval)
    movedata3[,currentcol] = replot(newtimepoints)
    currentcol = currentcol + 1
  }
  
  movedata4 = gather(movedata3, Stream, Value, -Time)
  movedata5 = movedata4 %>% separate(Stream, cleancolnames, sep = "#")
  
  if (length(unique(movedata5$Time))>length(unique(data.df$Time)))
  {
    warning("You have oversampled your data, meaning that you now have more datapoints than you originally collected. This may be misleading, use with care.")
  }
  
  movedata6 = vascr_remove_metadata(movedata5)
  movedata6 = vascr_explode(movedata6)
  
  return(movedata6)
  
}
Subset a continuous variable}
\usage{
vascr_subset_continuous(
  data,
  continuous,
  exact_match = FALSE,
  strip_empty = TRUE,
  implode = TRUE
)
}
\arguments{
\item{data}{A vascr dataset to subset}

\item{continuous}{The continuous variable to subset}

\item{exact_match}{Should the variables selected have an exact match to the column names input, or only contain the value input. Default is containing as otherwise you have to keep track of units.}

\item{strip_empty}{Should columns in which none of the selected variables are present be removed}

\item{implode}{Should the final data set be imploded, replacing the sample wells present}
}
\value{
A vascr dataset subsampled on a continuous variable
}
\description{
Subset the columns that are exploded out of a continuous variable. Contains options to remove descriptors that are now defunct, so this will be repaired later.
}
\examples{
# Sub code for breaking out continuous datasets
#exploded = vascr_explode(xcell)
#subset = vascr_subset_continuous(exploded, continuous = "ATP", strip_empty = FALSE)
#vascr_plot(exploded, unit = "CI", frequency = "10000", replication = "experiments"
# , normtime = 160, continuouscontains = "ATP")


}
\keyword{internal}
