% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_summarise}
\alias{vascr_summarise}
\title{Title}
\usage{
vascr_summarise(data.df, level = "wells")
}
\arguments{
\item{data.df}{}

\item{level}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_summarise(growth.df, level = "summary")

}
