% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_summarise_cross_correlation}
\alias{vascr_summarise_cross_correlation}
\title{Generate all cross-correlation pairs between sets of values}
\usage{
vascr_summarise_cross_correlation(
  data.df,
  reference = NULL,
  comparator = NULL,
  manualpairs = NULL
)
}
\arguments{
\item{data.df}{The vascr dataset to summarise}
}
\value{
A dataset contining values from each cross correlation pair
}
\description{
Generate all cross-correlation pairs between sets of values
}
\keyword{internal}
