% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_summarise_errortype}
\alias{vascr_summarise_errortype}
\title{Title}
\usage{
vascr_summarise_errortype(data.df, errortype)
}
\arguments{
\item{data.df}{}

\item{errortype}{}
}
\value{

}
\description{
Title
}
