% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_summarise_mean}
\alias{vascr_summarise_mean}
\title{Create the means from differnet wells, experiments or an overall summary}
\usage{
vascr_summarise_mean(data.df, level = "summary")
}
\arguments{
\item{data.df}{An ECIS dataset in standard format}

\item{level}{The level of replication to generate the summary at. Options are "experiment" or "summary"}
}
\value{
An ECIS dataset supplimented with summary statistics
}
\description{
Creates and ECIS dataset that has had all samples of the same type averaged together. Assumes that each sample is independent, IE that this function has already been run on individual experiments
}
\examples{

vascr_summarise_mean(growth.df, "summary")
vascr_summarise_mean(growth.df, "experiments")
vascr_summarise_mean(growth.df, "wells")

vascr_test_summary_level(growth.df)

}
