% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tests.R
\name{vascr_test_design}
\alias{vascr_test_design}
\title{Test the integrity of an ECIS dataframe}
\usage{
vascr_test_design(data.df, verbose = FALSE)
}
\arguments{
\item{data.df}{An ECIS dataframe, subset if needed}

\item{verbose}{How verbose to be. FALSE returns a dataset, TRUE returns a full verbose output of each test}
}
\value{
A whole slew of tests
}
\description{
This function will run a whole suite of tests on an ECIS dataframe, to check that it is both well designed for statistical analysis as well as technically intact
}
\examples{
# Prep a known good compliment of data
#data = vascr_subset(growth.df, time = 100, unit = "R", frequency = 4000)

# Test a full compliment of data passes
#vascr_test_design(data)

#vascr_test_design(growth.df)
# Test function picks up unbalanced replicate #'s
#data2 = vascr_exclude(data, well = c("A1", "B1", "C1"))
#vascr_test_design(data2)
# Test function picks up missing pairs (in this case due to Rb not being established)
#data3 = vascr_subset(growth.df, time = 100, unit = "Rb")
#vascr_test_design(data3)

}
\keyword{internal}
