% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{vascr_test_explosion_integrity}
\alias{vascr_test_explosion_integrity}
\title{Test that the datset is correctly formatted for explosion}
\usage{
vascr_test_explosion_integrity(data)
}
\arguments{
\item{data}{An ECIS data frame}
}
\value{
A boolean value
}
\description{
Test that the datset is correctly formatted for explosion
}
\examples{
#library(stringr)

#data.df = growth.df
#data.df$Sample = str_replace(data.df$Sample, " ", "_")
#data.df$Sample = paste(data.df$Sample, "+ 10_nm nothing important")
#data.df$Sample = paste(data.df$Sample, "+ 4_nm Carpet + ECV_line")

#vascr_test_explosion_integrity(growth.df)

}
