% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_titles}
\alias{vascr_titles}
\title{Generate human readable versions of the unit variable for graphing}
\usage{
vascr_titles(unit, frequency = 0, prefix = "")
}
\arguments{
\item{unit}{The unit to submit}

\item{frequency}{The frequency to submit}
}
\value{
An expression containing the correct data label for the unit
}
\description{
Generate human readable versions of the unit variable for graphing
}
\examples{

#vascr_titles("Rb")
#vascr_titles("R")


}
\keyword{internal}
