% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_titles_vector}
\alias{vascr_titles_vector}
\title{Title}
\usage{
vascr_titles_vector(units)
}
\arguments{
\item{units}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_titles_vector(c("Rb", "R", "Cm"))

}
