% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_tukey}
\alias{vascr_tukey}
\title{Generate a tukey analyasis of vascr data}
\usage{
vascr_tukey(data.df, unit, frequency, time, priority = NULL, raw = FALSE)
}
\arguments{
\item{data.df}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}

\item{priority}{}

\item{raw}{}
}
\value{

}
\description{
Generate a tukey analyasis of vascr data
}
\examples{
vascr_tukey(growth.df, "R", 4000, 100)
vascr_tukey(growth.df, "R", 4000, 100, raw = TRUE)

}
