% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_units_table}
\alias{vascr_units_table}
\title{Table of units used in the vascr package}
\usage{
vascr_units_table()
}
\value{
A data frame of units, their content and if they are modeled
}
\description{
Table of units used in the vascr package
}
\examples{
#vascr_units_table()

}
