% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excelligence.R
\name{xcelligence_import_exported}
\alias{xcelligence_import_exported}
\title{Import excelligence data files}
\usage{
xcelligence_import_exported(file, key)
}
\arguments{
\item{file}{Path to the excelligence file to be imported}

\item{key}{Optional, standard keyfile that will overwrite what is stored in the excelligence file}
}
\value{
A standard impedr data frame
}
\description{
Import excelligence data files
}
\examples{
#filename = "inst/extdata/xcell.txt"
#dataset = xcelligence_import_exported(filename)
#ecis_plot(dataset, unit = "CI", frequency = 0 , replication = "experiments", time = c(170,220) , samplecontains = "Nerifollin", normtime = 185, preprocessed = TRUE)

}
