% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_new_xcelligence.R
\name{xcelligence_import_generate_CI}
\alias{xcelligence_import_generate_CI}
\title{Generate CI from xcelligence data}
\usage{
xcelligence_import_generate_CI(data.df)
}
\arguments{
\item{data.df}{The dataset to generate CI from}
}
\value{
An enlargened dataset
}
\description{
Generate CI from xcelligence data
}
\keyword{internal}
