% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xcelligence.R
\name{xcelligence_lengthen_platemap}
\alias{xcelligence_lengthen_platemap}
\title{library(RODBC)
library(svSocket)
Lengthen out an excelligence platemap}
\usage{
xcelligence_lengthen_platemap(data)
}
\arguments{
\item{data}{The raw xcelligence dataset to deal with}
}
\value{
A slighlty tidier dataset
}
\description{
Switches the xcelligence format of having each column as a column and each row as a row, then stacking timepoints to a standardised tidy data format
}
\keyword{internal}
