% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{categorical_mode}
\alias{categorical_mode}
\title{Find the mode of a categorical variable}
\usage{
categorical_mode(x)
}
\arguments{
\item{x}{vector to find mode of}
}
\value{
the most commonly occouring character
}
\description{
Find the mode of a categorical variable
}
