% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{categorical_mode}
\alias{categorical_mode}
\title{#' Fix typographical errors in an ECIS dataframe
#' 
#' This function will go through all the sample and experiment names and replace all the values that are incorrect. Usefull for fixing up typographical errors, or places where you have named things differently inadvertently.
#'
#' @param data.df the data to correct
#' @param incorrect the error to be corrected
#' @param correct the corrrect string
#' @param limit can be set to "Sample or "Experiment" to limit the replacement to a single variable
#'
#' @return the repared dataframe
#' @export
#'
#' @examples
#' 
#' #vascr_fix_error(growth.df, "cells", "cells plated")
#' 
vascr_fix_error = function (data.df, incorrect, correct, limit = "None")
{
  
  if (limit == "None" || limit == "Sample")
  {
  data.df$Sample = gsub(incorrect, correct, data.df$Sample)
  }
  
  if (limit == "None" || limit == "Experiment")
  {
  data.df$Experiment = gsub(incorrect, correct, data.df$Experiment)
  }
  
  return(data.df)
}
Find the mode of a categorical variable}
\usage{
categorical_mode(x)
}
\arguments{
\item{x}{vector to find mode of}
}
\value{
the most commonly occouring character
}
\description{
#' Fix typographical errors in an ECIS dataframe
#' 
#' This function will go through all the sample and experiment names and replace all the values that are incorrect. Usefull for fixing up typographical errors, or places where you have named things differently inadvertently.
#'
#' @param data.df the data to correct
#' @param incorrect the error to be corrected
#' @param correct the corrrect string
#' @param limit can be set to "Sample or "Experiment" to limit the replacement to a single variable
#'
#' @return the repared dataframe
#' @export
#'
#' @examples
#' 
#' #vascr_fix_error(growth.df, "cells", "cells plated")
#' 
vascr_fix_error = function (data.df, incorrect, correct, limit = "None")
{
  
  if (limit == "None" || limit == "Sample")
  {
  data.df$Sample = gsub(incorrect, correct, data.df$Sample)
  }
  
  if (limit == "None" || limit == "Experiment")
  {
  data.df$Experiment = gsub(incorrect, correct, data.df$Experiment)
  }
  
  return(data.df)
}
Find the mode of a categorical variable
}
\examples{
categorical_mode(c("Cat", "Cat", "Monkey"))

}
