% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cellzscope.R
\name{cellzscope_import}
\alias{cellzscope_import}
\title{Import cellSZcope data}
\usage{
cellzscope_import(raw, model, key)
}
\arguments{
\item{raw}{File location of the raw dataset}

\item{model}{File locaiton of the modeled dataset}

\item{key}{Location of a vascar standard lookup table. Optional, but can be used to import more granular data than possible with the built in row names}
}
\value{
a standard vascar dataset
}
\description{
Data must first be exported in two parts from the cellZScope software. This can then be imported to a standard vascar dataset with this function
}
\examples{



}
