% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cellzscope.R
\name{cellzscope_import_raw}
\alias{cellzscope_import_raw}
\title{Import a raw cellZScope dataset}
\usage{
cellzscope_import_raw(raw, key, experimentname)
}
\arguments{
\item{raw}{the locaiton of spectral data exported from the cellZscope software.}

\item{key}{Optional, allows for sample names to be assigned to wells}

\item{experimentname}{Name of the experiment to be built into the dataset}
}
\value{
A vascar compatable dataset
}
\description{
Data must first be exported from the cellZscope software.
}
\examples{
#raw = system.file("extdata/mdckspectra.txt", package = "vascr")
#key = system.file("extdata/mdckkey.csv", package = "vascr")

#output = cellzscope_import_raw(raw, key)
#output = vascr_subset(output, time = c(0,50))
#vascr_plot(output, unit = "R", frequency = 4000, replication = "wells")
}
\keyword{internal}
