% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cellzscope.R
\name{cellzscope_import_raw}
\alias{cellzscope_import_raw}
\title{Import a raw cellZScope dataset}
\usage{
cellzscope_import_raw(raw)
}
\arguments{
\item{raw}{the locaiton of spectral data exported from the cellZscope software.}
}
\value{
A vascar compatable dataset
}
\description{
Data must first be exported from the cellZscope software.
}
\examples{



}
