% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{do.call_relevant}
\alias{do.call_relevant}
\title{Title}
\usage{
do.call_relevant(name, payload, arguments)
}
\arguments{
\item{name}{}

\item{payload}{}

\item{arguments}{}
}
\value{

}
\description{
Title
}
