% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ecis_combine}
\alias{ecis_combine}
\title{Combine ECIS data frames end to end}
\usage{
ecis_combine(...)
}
\arguments{
\item{...}{List of data frames to be combined}
}
\value{
A single data frame containing all the data imported, automaticaly incremented by experiment
}
\description{
This funciton will combine ECIS datasets end to end. Preferential to use over a simple rbind command as it runs additional checks to ensure that datapoints are correctly generated
}
\examples{

#Make two fake experiments worth of data

experiment1.df = ecis_subset(growth.df, experiment = "1")
experiment2.df = ecis_subset(growth.df, experiment = "2")
experiment3.df = ecis_subset(growth.df, experiment = "3")

ecis_combine(experiment1.df, experiment2.df, experiment3.df)

}
