% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmaster.R
\name{ecis_combine}
\alias{ecis_combine}
\title{Combine data frames end to end}
\usage{
ecis_combine(...)
}
\arguments{
\item{...}{List of data frames to be combined}
}
\value{
A single data frame containing all the data imported, automaticaly incremented by experiment
}
\description{
Combine data frames end to end
}
\examples{

#Make two fake experiments worth of data

experiment1.df = data.df
experiment2.df = data.df

ecis_combine(experiment1.df, experiment2.df)

}
