% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{ecis_explode}
\alias{ecis_explode}
\title{Explode out continuous data from Sample}
\usage{
ecis_explode(data)
}
\arguments{
\item{data}{Standard ECIS data frame to explode}
}
\value{
The data frame to return
}
\description{
Explode out continuous data from Sample
}
\examples{
library(stringr) # Needed to generate the dummy data

data.df = growth.df
data.df$Sample = str_replace(data.df$Sample, " ", "_")
data.df$Sample = paste(data.df$Sample, "+ 10_nm nothing important")
data.df$Sample = paste(data.df$Sample, "+ 4_nm Carpet + ECV_line")
processed = ecis_explode(data.df)
}
