% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_generate_continuous}
\alias{ecis_generate_continuous}
\title{Generate CSV separated continuous variables}
\usage{
ecis_generate_continuous(string)
}
\arguments{
\item{string}{The string to be converted to a continuous vector}
}
\value{
A CSV separated list of continuous variables
}
\description{
Generate CSV separated continuous variables
}
\examples{

ecis_generate_continuous("  30,000.01 cells + 1 mg/ml TNFa")

}
