% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ecis.R
\name{ecis_import}
\alias{ecis_import}
\title{Import all ECIS values, a child of ecis_import_raw and ecis_import_model}
\usage{
ecis_import(rawdata, modeled, key, experimentname = "NA")
}
\arguments{
\item{rawdata}{A raw ABP file to import}

\item{modeled}{A modeled APB file for import}

\item{key}{A CSV file containing each well and it's corresponding sample}

\item{experimentname}{Name of the experiment to be built into the dataset}
}
\value{
A dataframe containing all the data APB generated from an experiment
}
\description{
Import all ECIS values, a child of ecis_import_raw and ecis_import_model
}
\examples{

#rawdata = system.file('extdata/growth1_raw_TimeResample.abp', package = 'vascr')
#modeled = system.file('extdata/growth1_raw_TimeResample_RbA.csv', package = 'vascr')
#key = system.file('extdata/growth1_samples.csv', package = 'vascr')

#Then run the import

#data = ecis_import(rawdata,modeled,key, experimentname = "TEST")
#head(data)
}
