% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ecis_import}
\alias{ecis_import}
\title{Import all ECIS values, a child of ecis_import_raw and ecis_import_model}
\usage{
ecis_import(resample, modeled, key, by, from = Inf, to = Inf,
  zero_time = 0, verbose = TRUE, experiment = NULL)
}
\arguments{
\item{resample}{A resampled APB file for import}

\item{modeled}{A modeled APB file for import}

\item{key}{A CSV file containing each well and it's corresponding sample}
}
\value{
A dataframe containing all the data APB generated from an experiment
}
\description{
Import all ECIS values, a child of ecis_import_raw and ecis_import_model
}
\examples{

location_of_resampled_data = system.file('Resample.abp', package = 'ecisr')
location_of_modeled_data = system.file('Model.csv', package = 'ecisr')
location_of_sample_defintions = system.file('Samples.csv', package = 'ecisr')

#Then run the import

data = ecis_import(location_of_resampled_data,location_of_modeled_data,location_of_sample_defintions, 5, 0, 200)
head(data)

}
