% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ecis.R
\name{ecis_import_model}
\alias{ecis_import_model}
\title{Import raw modeled data}
\usage{
ecis_import_model(modeleddata, sampledefine, experimentname)
}
\arguments{
\item{modeleddata}{Raw modeled data in APB format}

\item{sampledefine}{CSV file containing which wells correspond to which values}
}
\value{
Data frame containing modeled data
}
\description{
Import raw modeled data
}
\examples{

 #First determine the locatins of your files relative to your dataset.
 #Here we use system.file to pull a default out of the filesystem, but you 
 #can use a path relative to the file you are working on. 
 #E.G 'Experiment1/Raw.abp'

modeleddata = system.file('extdata/growth1_raw_TimeResample_RbA.csv', package = 'vascr')
sampledefine = system.file('extdata/growth1_samples.csv', package = 'vascr')

#Then run the import

data = ecis_import_model(modeleddata, sampledefine)
head(data)
vascr_plot(data, unit = "Rb", replication = "wells")

}
