% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmaster.R
\name{ecis_import_model_long}
\alias{ecis_import_model_long}
\title{Title}
\usage{
ecis_import_model_long(rawdata, samples)
}
\arguments{
\item{rawdata}{Raw modeled data in APB format}

\item{samples}{CSV file containing which wells correspond to which values}
}
\value{
Data frame containing modeled data
}
\description{
Title
}
\examples{

#' #First determine the locatins of your files relative to your dataset. Here we use system.file to pull a default out of the filesystem, but you can use a path relative to the file you are working on. E.G "Experiment1/Raw.abp"

location_of_modeled_data = system.file("Model.csv", package = "ECISR")
location_of_sample_defintions = system.file("Samples.csv", package = "ECISR")

#Then run the import

ecis_import_model_long(location_of_modeled_data, location_of_sample_defintions)

}
