% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ecis_import_raw}
\alias{ecis_import_raw}
\title{ECIS raw data importer}
\usage{
ecis_import_raw(rawdata, sampledefine)
}
\arguments{
\item{rawdata}{A resampled ABP file containing the un-modeled data}

\item{sampledefine}{A CSV file containing well numbers and their corresponding sample names}
}
\value{
Data frame containing all the raw data readings from the ECIS Z0 instrument
}
\description{
Raw data importer, generates a r dataframe from a raw ABP file
}
\examples{

#First determine the locatins of your files relative to your dataset. 
#Here we use system.file to pull a default out of the filesystem, 
#but you can use a path relative to the file you are working on. 
#E.G 'Experiment1/Raw.abp'

location_of_resampled_data = system.file('Resample.abp', package = 'ECISR')
location_of_sample_defintions = system.file('Samples.csv', package = 'ECISR')

#Then run the import

ecis_import_raw(location_of_resampled_data, location_of_sample_defintions)

}
