% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plot}
\alias{ecis_plot}
\title{ECIS plot}
\usage{
ecis_plot(data, unit, frequency = 0, replication = "all", time = Inf,
  samplesubset = "")
}
\arguments{
\item{data}{A standard ECIS data frame to plot}

\item{unit}{Unit to plot}

\item{frequency}{The frequency of data to display. All modelled variables have a frequency of 0}

\item{replication}{How much of the replicaiton to display. Options are 'all', 'experiment', 'summary'.}

\item{time}{The time to subset if a slice is required. If set to Inf all data will be displayed}

\item{samplesubset}{Optional, only samples that contain this string will be plotted. Standard search and wildcard searches apply.}
}
\value{
A ggplot2 object
}
\description{
ECIS plot
}
\examples{
ecis_plot(data.df, 'Rb', 0, 'all')
ecis_plot(data.df, 'R', 4000, 'summary', time = 75)


}
