% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_plot}
\alias{ecis_plot}
\title{ECIS plot}
\usage{
ecis_plot(data, unit = "R", frequency = 4000,
  replication = "summary", time = Inf, samplecontains = "",
  experiment = "", error = 1, linesize = 1, errorsize = 1,
  alphavalue = 0.1, confidence = 1, xlab = "Time (hours)",
  ylab = "Value", title = "Title")
}
\arguments{
\item{data}{A standard ECIS data frame to plot}

\item{unit}{Unit to plot}

\item{frequency}{The frequency of data to display. All modelled variables have a frequency of 0}

\item{replication}{How much of the replicaiton to display. Options are 'all', 'experiment', 'summary'.}

\item{time}{The time to subset if a slice is required. If set to Inf all data will be displayed}

\item{samplecontains}{Optional, only samples that contain this string will be plotted. Standard search and wildcard}

\item{experiment}{Optional, allows you to limit the experiment plotted. Experiment names should be separated with |
searches apply.}

\item{error}{Display error bars. 0 displays no error bars, 1 displays them all. Higher numbers will reduce the frequency of error bars plotted.}

\item{linesize}{Width of mean lines shown on graphs}

\item{errorsize}{Width of error bars shown on graphs}

\item{alphavalue}{Alpha value of area enclosed by error bars. May be lowered for buisy graphs}

\item{confidence}{The confidence to use when generating basic significance plots}

\item{xlab}{X axis value}

\item{ylab}{Y axis value}

\item{title}{The title of the ggplot}
}
\value{
A ggplot2 object
}
\description{
The master function for generating ggplot2 objects from ECIS data. Incorporates all the summarising, testing and graphing operations into a single function.
}
\examples{
ecis_plot(growth.df, 'Rb', replication = 'summary', 
error = 2, linesize = 1, errorsize = 1, alphavalue = .1, title = "Cars", xlab = "Hours")
ecis_plot(growth.df, 'Rb', replication = 'all',
 error = 2, linesize = .1, errorsize = 1, alphavalue = .1, title = "Cars", xlab = "Hours")
 ecis_plot(growth.df, 'Rb', replication = 'experiment',
 error = 2, linesize = .1, errorsize = 1, alphavalue = .1, title = "Cars", ylab = "Rb"
 , xlab = "Hours")
ecis_plot(growth.df, 'R', 4000, 'summary', time = 75)
ecis_plot(growth.df, "R", "4000", "summary", 50, confidence = 0.1)

}
