% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_plot_isolate}
\alias{ecis_plot_isolate}
\title{Graph a single well that is misbehaving}
\usage{
ecis_plot_isolate(data.df, well, title = "Well isolation plot",
  unit = "R", frequency = 4000, ...)
}
\arguments{
\item{data.df}{The dataset the well is in}

\item{well}{The well to be isolated}

\item{title}{The title of the plot}

\item{unit}{The unit to plot. Default is R}

\item{frequency}{The frequency to plot. Default is 4000}

\item{...}{Other conditions to pass on to the ecis_plot command that generates the graph}
}
\value{
A ggplot graph showing the isoated well and the experimental median well
}
\description{
Graph a single well that is misbehaving
}
\examples{

ecis_plot_isolate(growth.df, well = "A3")

}
