% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{ecis_plot_plate}
\alias{ecis_plot_plate}
\title{Plot the wells of a 96 well plate}
\usage{
ecis_plot_plate(data.df, unit = "R", frequency = 4000)
}
\arguments{
\item{data.df}{A standard ECIS dataframe. Ideally this will contain only one experiment, but multiple experiments are supported.}

\item{unit}{The unit to plot. Default is R}

\item{frequency}{The frequency to plot. Default is 4000}
}
\value{
A GGplot2 matrix
}
\description{
Plot the wells of a 96 well plate
}
\examples{
ecis_plot_plate(growth.df, unit = "Rb")

}
