% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{ecis_plotvariable}
\alias{ecis_plotvariable}
\title{Generate a standard ECIS data representation}
\usage{
ecis_plotvariable(data.df, unit, frequency)
}
\arguments{
\item{data.df}{An ECIS data set. Must be}

\item{unit}{The unit to plot}

\item{frequency}{Frequency to plot. All modeled units have a default frequency of 0}
}
\value{
A GGplot2 object
}
\description{
THIS FUNCTION IS DEPRECIATED, Use ECIS_plot_xxx series of funcitons for finer control over how to deal with replicate experiments. This graphic will underestimate error from replicate experiments.
}
\examples{
ecis_plotvariable(data.df, "Rb", 0)

}
