% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{ecis_prep_graphdata}
\alias{ecis_prep_graphdata}
\title{Title}
\usage{
ecis_prep_graphdata(
  data,
  unit,
  frequency,
  time,
  samplecontains,
  experiment,
  error,
  alignkey,
  normtime,
  divide,
  preprocessed,
  continuouscontains,
  stripidentical
)
}
\arguments{
\item{data}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}

\item{samplecontains}{}

\item{experiment}{}

\item{error}{}

\item{alignkey}{}

\item{normtime}{}

\item{divide}{}

\item{preprocessed}{}

\item{continuouscontains}{}

\item{stripidentical}{}
}
\value{

}
\description{
Title
}
