% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_methods.R
\name{ecis_roundtime}
\alias{ecis_roundtime}
\title{Title}
\usage{
ecis_roundtime(data.df, time)
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{time}{The time point that needs rounding}
}
\value{
A timepoint that exactly aligns with a measured datapoint
}
\description{
Title
}
