% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecis_roundtime}
\alias{ecis_roundtime}
\title{Standardise how well numbers are represented}
\usage{
ecis_roundtime(data.df, time)
}
\arguments{
\item{data.df}{A standard ECIS data frame}

\item{time}{The time point that needs rounding}
}
\value{
A timepoint that exactly aligns with a measured datapoint
}
\description{
Standardise how well numbers are represented
}
\examples{
ecis_roundtime(growth.df, 100)

}
