% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_master.R
\name{ecis_standardise_wells}
\alias{ecis_standardise_wells}
\title{Standardise well names accross import types}
\usage{
ecis_standardise_wells(well)
}
\arguments{
\item{well}{The well to be standardised}
}
\value{
Standardised well names
}
\description{
Replaces A01 in strings with A0. Important for importing ABP files which may use either notation.
}
\examples{
ecis_standardise_wells('A01')

}
