% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{ecis_subset}
\alias{ecis_subset}
\title{Downsample data}
\usage{
ecis_subset(data.df, nth)
}
\arguments{
\item{data.df}{An ECIS dataset}

\item{nth}{An integer. Every nth value will be preserved in the subsetting}
}
\value{
Downsampled ECIS data set
}
\description{
Returns a subset of the original data set that has only every nth value. Greatly increases computational preformance for a minimal loss in resolution during time course experiments.
}
\examples{

ecis_subset(data.df, 50)

}
