% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ecis_subset}
\alias{ecis_subset}
\title{Subset an ECIS dataset on multiple factors}
\usage{
ecis_subset(data.df, time = Inf, unit = "", frequency = Inf,
  samplecontains = "", experiment = "", well = "")
}
\arguments{
\item{data.df}{A standard ECIS dataset}

\item{time}{The time to subset at. Default will line plot all data, can also submit a vector of length 2
and the times between those two points will be submited.}

\item{unit}{The unit requred}

\item{frequency}{The frequency at which the reading was taken. All modeled variables have a frequency of 0}

\item{samplecontains}{The samples to plot. A string that is searched accross all sample names, and those that match are plotted.}

\item{experiment}{The experiment to plot. Default is all experiments}

\item{well}{The wells required}
}
\value{
A smaller ECIS dataset
}
\description{
Generates a cut down dataset for processing purposes. Used heavily by all other internal functions, but may also be useful for inspecting digestable chunks of raw data.
}
\examples{
data = ecis_subset(growth.df, time = c(20.23,50.73), frequency = 4000, unit = "R", 
samplecontains = "05,000", experiment = "2", well = "G5")
head(data)
data = ecis_subset(growth.df, time = c(20.23,50.73), frequency = 4000, unit = "R", 
samplecontains = "05,000", experiment = "2")
head(data)

}
