% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{ecis_test_exploded}
\alias{ecis_test_exploded}
\title{Test if an ECIS data frame is exploded}
\usage{
ecis_test_exploded(data)
}
\arguments{
\item{data}{the data frame to test}
}
\value{
True (exploded) or false (not exploded)
}
\description{
This function tests to see if exploded columns are present in a dataset. Does not check that all are present, or that they are corrrectly exploded as this would be much slower.
}
\examples{
ecis_test_exploded(growth.df)
imploded = ecis_implode(growth.df)
ecis_test_exploded(imploded)

}
