% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecisgraphics.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Internal funciton for aligning grids}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{GGPlot objects to combine}

\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{position}{Where to place each graph. Unsure about this one.}
}
\value{
A multi-plot graph object
}
\description{
Internal funciton for aligning grids
}
\examples{

#graph1 = ecis_plot(data.df, 'Rb', 0, 'all')
#graph2 = ecis_plot(data.df, 'Rb', 0, 'experiment')
#graph3 = ecis_plot(data.df, 'Rb', 0, 'summary')

#grid_arrange_shared_legend(graph1, graph2, graph3, ncol = 1, nrow = 3)

}
