% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xcelligence.R
\name{import_xcelligence}
\alias{import_xcelligence}
\title{Import an xcelligence file}
\usage{
import_xcelligence(file, key, experimentname = "NA", password = "RTCaDaTa")
}
\arguments{
\item{file}{The file to import}

\item{key}{A keyfile to apply. Optional, as the xCELLigence internal definitions will be used if no file is specified}

\item{experimentname}{Name of the experiment to be built into the dataset}
}
\value{
A vascr datafile
}
\description{
Import an xcelligence file
}
\examples{
# xCELLigence test
rawdata = system.file('extdata/instruments/xcell.plt', package = 'vascr')
sampledefine = system.file('extdata/instruments/xcellkey.csv', package = 'vascr')

# data7 = import_xcelligence(file = rawdata, key = sampledefine, "TEST7")

 
}
\keyword{internal}
