% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_carry_down_names}
\alias{vascr_carry_down_names}
\title{Title}
\usage{
vascr_carry_down_names(data.df, cols_to_carry = NULL, remove_blank = TRUE)
}
\arguments{
\item{data.df}{}

\item{cols_to_carry}{}

\item{remove_blank}{}
}
\value{

}
\description{
Title
}
