% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_cross_correlation.R
\name{vascr_cc}
\alias{vascr_cc}
\title{Title}
\usage{
vascr_cc(data.df, reference = "none")
}
\description{
Title
}
\examples{
data.df = vascr::growth.df \%>\% vascr_subset(unit = "R", frequency = 4000, sampleid = c(1,4,7))

vascr_cc(data.df)

vascr_cc(data.df, reference = "35,000_cells + HCMEC D3_line")
vascr_cc(data.df, reference = c("35,000_cells + HCMEC D3_line", "5,000_cells + HCMEC D3_line"))

}
