% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_col_exists}
\alias{vascr_col_exists}
\title{Check if a column exists in a dataset}
\usage{
vascr_col_exists(col, data.df)
}
\arguments{
\item{col}{Column to search for}

\item{data.df}{Dataset to search}
}
\value{
A boolean
}
\description{
Check if a column exists in a dataset
}
\examples{
# vascr_col_exists("Well", data.df)
}
\keyword{internal}
