% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova.R
\name{vascr_dunnett}
\alias{vascr_dunnett}
\title{Title}
\usage{
vascr_dunnett(data.df, unit, frequency, time, reference)
}
\arguments{
\item{data.df}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}

\item{reference}{}
}
\description{
Title
}
\examples{
vascr_dunnett(growth.df, "R", 4000, 100, 6)

}
