% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_subset.R
\name{vascr_exclude}
\alias{vascr_exclude}
\title{Exclude samples from a vascr dataset}
\usage{
vascr_exclude(data.df, well = NULL, experiment = NULL, sampleid = NULL)
}
\arguments{
\item{data.df}{the vascr data set to exclude things from}

\item{well}{wells to exclude}

\item{experiment}{experiments to exclude}

\item{sampleid}{sampleID (or vector or sampleIDs) to exclude from analysis}
}
\value{
A smaller vascr dataset
}
\description{
Exclude samples from a vascr dataset
}
\examples{
vascr_exclude(growth.df, c("A01", "E01"))
vascr_exclude(growth.df, sampleid = 1)


}
