% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_cont.R
\name{vascr_explode}
\alias{vascr_explode}
\title{Explode out continuous data from Sample}
\usage{
vascr_explode(data)
}
\arguments{
\item{data}{Standard ECIS data frame to explode}
}
\value{
The data frame to return
}
\description{
Explode out continuous data from Sample
}
\examples{
growth.df$Instrument ="ECIS"

# Stip out all the non-core columns in the dataset
data.df = vascr_remove_metadata(growth.df)

# Run the explosion to re-generate the non-core columns
processed = vascr_explode(data.df)

# Show that the re-created data is identical to the original dataset
all.equal(growth.df, processed)

}
