% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_explode_wells}
\alias{vascr_explode_wells}
\title{Explode the wells in a VASCR dataset}
\usage{
vascr_explode_wells(data.df, separate_rows = FALSE)
}
\arguments{
\item{data.df}{The dataset to explode}

\item{separate_rows}{Split cells onto multiple rows if wells such as "A1,A2" may be present in the dataset}
}
\value{
A vascr dataset with rows and columns exploded
}
\description{
Tools for exploding wells out into row and column variables, and separating comma separated well values if needed.
}
\examples{
#vascr_explode_wells(growth.df)
#vascr_explode_wells(growth.df, separate_rows = TRUE)

}
\keyword{internal}
