% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_export.R
\name{vascr_export}
\alias{vascr_export}
\title{Export a vascr dataframe}
\usage{
vascr_export(data.df, filepath = NULL)
}
\arguments{
\item{data.df}{a vascr dataset to export}

\item{filepath}{Path to save the dataframe in}
}
\value{
A dataframe in prism format, or writes to file if filepath specified
}
\description{
Export a vascr dataframe
}
\examples{
filepath = tempfile("test_export", fileext = ".xlsx")
vascr_export(growth.df \%>\% vascr_subset(unit = c("R", "Rb"), frequency = c(0,4000)), filepath)

}
