% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_find.R
\name{vascr_find}
\alias{vascr_find}
\title{Title}
\usage{
vascr_find(data.df, paramater, value = NA)
}
\arguments{
\item{data.df}{}

\item{paramater}{}
}
\description{
Title
}
\examples{

vascr_find(growth.df, "Time")
vascr_find(growth.df, "Time", 66.97)

vascr_find(growth.df, "Unit")
vascr_find(growth.df, "Unit", "Rb")

vascr_find(growth.df, "Well")
vascr_find(growth.df, "Well", "A1")

vascr_find(growth.df, "Sample")
vascr_find(growth.df, "Sample", "5000 cells")

vascr_find(growth.df, "Frequency")
vascr_find(growth.df, "Frequency")

vascr_find(growth.df, "Experiment")
vascr_find(growth.df, "Experiment")

vascr_find(growth.df, "SampleID")
vascr_find(growth.df, "SampleID")
}
