% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_find_sample}
\alias{vascr_find_sample}
\title{Find vascr samples that meet the criteria}
\usage{
vascr_find_sample(
  data.df,
  sample_contains = NULL,
  sample_not_contains = NULL,
  sample_equal_to = NULL,
  variable_set = NULL,
  variable_unset = NULL,
  variable_equal_to = NULL,
  variable_not_equal_to = NULL,
  variable_greater_than = NULL,
  variable_less_than = NULL,
  include_vehicle = TRUE
)
}
\arguments{
\item{data.df}{The dataset to screen}

\item{sample_contains}{Returns that samples that contain a string. This is very broad.}

\item{sample_not_contains}{}

\item{sample_equal_to}{}

\item{variable_set}{}

\item{variable_unset}{}

\item{variable_equal_to}{}

\item{variable_not_equal_to}{}

\item{variable_greater_than}{}

\item{variable_less_than}{}

\item{include_vehicle}{}
}
\value{

}
\description{
Search the samples, based on pairs of datasets. Pass to any variable names of variables paired with what is beng screened for. Using include_vehicle will respect variable
}
\examples{

vascr_find_sample(xcell, variable_set = "ATP")
vascr_find_sample(growth.df)

vascr_find_sample(czs)

data.df = vascr_subset(unifiedr, unit = "Rb")
vascr_find_sample(data.df, variable_equal_to = c("HCMVEC", 20000))

}
