% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_find.R
\name{vascr_find_sample}
\alias{vascr_find_sample}
\title{Title}
\usage{
vascr_find_sample(data.df, sample)
}
\arguments{
\item{data.df}{}

\item{sample}{}
}
\description{
Title
}
\examples{

sample = 3000

vascr_find_sample(growth.df, sample)

}
