% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_force_median}
\alias{vascr_force_median}
\title{Title}
\usage{
vascr_force_median(vector, round = "up")
}
\arguments{
\item{vector}{}

\item{round}{}
}
\value{

}
\description{
Title
}
\examples{
vector = unique(growth.df$Frequency)
vascr_force_median(vector)
}
