% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_full_implode}
\alias{vascr_full_implode}
\title{Implode multiple columns from a data frame into a single title}
\usage{
vascr_full_implode(data.df, cols_to_implode = NULL, remove_blank = TRUE)
}
\arguments{
\item{data.df}{The dataset to implode}

\item{cols_to_implode}{The columns to implode}

\item{remove_blank}{Should blank ID's be removed}
}
\value{
A data frame
}
\description{
Implode multiple columns from a data frame into a single title
}
\examples{
vascr_full_implode(growth.df \%>\% vascr_subset(unit = "Rb"))
vascr_full_implode(growth.df \%>\% vascr_subset(unit = "Rb"), c("cells", "Experiment"))

}
