% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_gg_color_hue}
\alias{vascr_gg_color_hue}
\title{Generate ggplot colour hues, for manually specifiying colours that match the default theme}
\usage{
vascr_gg_color_hue(n)
}
\arguments{
\item{n}{Number of variables to access}
}
\value{
A vector of ggplot colours
}
\description{
Generate ggplot colour hues, for manually specifiying colours that match the default theme
}
