% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_implode}
\alias{vascr_implode}
\title{Title}
\usage{
vascr_implode(
  data.df,
  stripidentical = TRUE,
  colnames = NULL,
  addcols = NULL,
  dropcols = NULL,
  fill_blank = "Control",
  cleanup_sample = FALSE
)
}
\arguments{
\item{data.df}{}
}
\description{
Title
}
