% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_naming.R
\name{vascr_implode}
\alias{vascr_implode}
\title{Implode individual samples from a vascr dataset}
\usage{
vascr_implode(data.df)
}
\arguments{
\item{data.df}{A vascr dataset to be imploded}
}
\value{
A vascr dataset with individual wells imploded
}
\description{
Implode individual samples from a vascr dataset
}
\examples{
vascr_implode(growth.df)

}
