% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_naming.R
\name{vascr_import}
\alias{vascr_import}
\title{Title}
\usage{
vascr_import(
  instrument = NULL,
  raw = NULL,
  modeled = NULL,
  map = NULL,
  experiment = NULL
)
}
\arguments{
\item{instrument}{}

\item{raw}{}

\item{modeled}{}

\item{map}{}

\item{experiment}{}
}
\description{
Title
}
