% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_interpolate_time}
\alias{vascr_interpolate_time}
\title{Title}
\usage{
vascr_interpolate_time(
  data.df,
  npoints = vascr_time_samples(data.df),
  from = min(floor(data.df$Time)),
  to = max(ceiling(data.df$Time))
)
}
\arguments{
\item{data.df}{}

\item{npoints}{}
}
\description{
Title
}
\examples{
# data.df = ecis1 \%>\% vascr_subset(time = c(1,2), unit = "R", frequency = 4000, well = c("D01", "D02", "D03"))
# from = 0.03613
# to = 0.5767

}
