% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_interpolate_time}
\alias{vascr_interpolate_time}
\title{Title}
\usage{
vascr_interpolate_time(
  data.df,
  npoints = vascr_time_samples(data.df),
  from = min(floor(data.df$Time)),
  to = max(ceiling(data.df$Time))
)
}
\arguments{
\item{data.df}{}

\item{npoints}{}
}
\value{

}
\description{
Title
}
