% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_is_modeled_unit}
\alias{vascr_is_modeled_unit}
\title{Check if a selected unit is modelled}
\usage{
vascr_is_modeled_unit(unit)
}
\arguments{
\item{unit}{The vascr symbol for the unit}
}
\value{
A boolean, true if it is modelled, false if it is raw electrical data
}
\description{
Check if a selected unit is modelled
}
\examples{
#vascr_is_modeled_unit("R")
#vascr_is_modeled_unit("Rb")
#vascr_is_modeled_unit(c("R", "Rb"))

}
