% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_is_valid_unit}
\alias{vascr_is_valid_unit}
\title{Test if a unit is valid}
\usage{
vascr_is_valid_unit(data.df, unit)
}
\arguments{
\item{data.df}{}

\item{unit}{}
}
\value{

}
\description{
Test if a unit is valid
}
\examples{
vascr_is_valid_unit(growth.df, "TEER")
vascr_is_valid_unit(growth.df, "TER")
vascr_is_valid_unit(growth.df, "Rb")

}
