% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_make_name}
\alias{vascr_make_name}
\title{#' Title
#'
#' @param data.df 
#' @param select_cols 
#' @param remove_blank 
#' @param fill_blank 
#'
#' @return
#' @export
#'
#' @examples
vascr_make_name = function(data.df, select_cols = NULL, remove_blank = TRUE, fill_blank = "Control")
{
  imploded = vascr_full_implode(data.df, cols_to_implode = NULL)
  return(imploded)
}}
\usage{
vascr_make_name(data.df, clean_output = TRUE, fill_blank = "Control")
}
\description{
#' Title
#'
#' @param data.df 
#' @param select_cols 
#' @param remove_blank 
#' @param fill_blank 
#'
#' @return
#' @export
#'
#' @examples
vascr_make_name = function(data.df, select_cols = NULL, remove_blank = TRUE, fill_blank = "Control")
{
  imploded = vascr_full_implode(data.df, cols_to_implode = NULL)
  return(imploded)
}
}
