% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecis_graphics.R
\name{vascr_make_panel}
\alias{vascr_make_panel}
\title{Title}
\usage{
vascr_make_panel(..., plots = NULL)
}
\arguments{
\item{...}{}

\item{plots}{}
}
\value{

}
\description{
Title
}
\examples{
plots[[1]] = vascr_plot(growth.df, "Rb")
plots[[2]] = vascr_plot(growth.df, "Cm")
vascr_make_panel(plots = plots)
vascr_make_panel(vascr_plot(growth.df, unit = "Cm"), vascr_plot(growth.df, unit = "Rb"))

}
