% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_match}
\alias{vascr_match}
\title{Match a string with the closest available option}
\usage{
vascr_match(tomatch, vector)
}
\arguments{
\item{tomatch}{The string to match}

\item{vector}{The vector to match into}
}
\value{
A character string of the closest matched string
}
\description{
Match a string with the closest available option
}
\examples{
# tomatch = "Rs"
# vector = vascr_find_unit(growth.df, "all")

# vascr_match(tomatch, vector)
# vascr_match("Rb", vector)
}
\keyword{internal}
