% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_next_in_vector}
\alias{vascr_next_in_vector}
\title{Title}
\usage{
vascr_next_in_vector(value, vector)
}
\arguments{
\item{value}{}

\item{vector}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_next_in_vector("Cm", vector)
vascr_next_in_vector("X", vector)
}
