% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_odd_in_vector}
\alias{vascr_odd_in_vector}
\title{Title}
\usage{
vascr_odd_in_vector(value, vector)
}
\arguments{
\item{value}{}

\item{vector}{}
}
\value{

}
\description{
Title
}
\examples{
vector = vascr_units_table()$Unit
vascr_odd_in_vector("Alpha", vector)
vascr_odd_in_vector("Cm", vector)
vascr_odd_in_vector("Drift", vector)


}
