% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_anova}
\alias{vascr_plot_anova}
\title{Make a display with all the anova analyasis pre-conducted}
\usage{
vascr_plot_anova(data.df, unit, frequency, time, priority = NULL, ...)
}
\arguments{
\item{data.df}{vascr dataset to plot}

\item{unit}{unit to plot}

\item{frequency}{frequnecy to plot}

\item{time}{timepoint to plot at}

\item{priority}{analyasis priority. Will use default if not filled in.}

\item{...}{other controls to be pushed to publish plot and generate graphdata}
}
\value{
A matrix of different ANOVA tests
}
\description{
Make a display with all the anova analyasis pre-conducted
}
\examples{

# vascr_plot_anova(growth.df, "R", 4000, 100)

}
