% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova.R
\name{vascr_plot_anova}
\alias{vascr_plot_anova}
\title{Make a display with all the anova analysis pre-conducted}
\usage{
vascr_plot_anova(data.df, unit, frequency, time, reference = NULL)
}
\arguments{
\item{data.df}{vascr dataset to plot}

\item{unit}{unit to plot}

\item{frequency}{frequnecy to plot}

\item{time}{timepoint to plot at}

\item{reference}{Sample to reference post-hoc analyasis to}
}
\value{
A matrix of different ANOVA tests
}
\description{
Make a display with all the anova analysis pre-conducted
}
\examples{

vascr_plot_anova(data.df = growth.df, unit = "R", frequency = 4000, time = 100)
vascr_plot_anova(data.df = growth.df, unit = "R", frequency = 4000, time = 100, 
          reference = "5,000_cells + HCMEC D3_line")

}
