% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_plot_anova_bar_reference}
\alias{vascr_plot_anova_bar_reference}
\title{Make a bar graph with ANOVA stars annotated against a set reference}
\usage{
vascr_plot_anova_bar_reference(
  data.df,
  unit,
  frequency,
  time,
  reference = min(data.df$SampleID),
  breaklines = TRUE
)
}
\arguments{
\item{data.df}{a vascr dataframe}

\item{unit}{the unit to use}

\item{frequency}{the frequency to use}

\item{time}{the time to use for the bar plot and ANOVA}

\item{reference}{SampleID of the sample to use as the reference for statistical analysis}
}
\description{
Make a bar graph with ANOVA stars annotated against a set reference
}
\examples{

vascr_plot_anova_bar_reference(growth.df, "R", 4000, 48)

}
