% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_bar_anova}
\alias{vascr_plot_bar_anova}
\title{Title}
\usage{
vascr_plot_bar_anova(data, priority, confidence, ...)
}
\arguments{
\item{data}{}

\item{confidence}{}

\item{...}{Any arguement to vascr_prep_graphdata or vascr_polish_plot}
}
\value{

}
\description{
Title
}
\examples{
vascr_plot_bar_anova(growth.df, confidence = 0.04999, unit = "Rb", time = 100, rotate_x_angle = 45)

}
