% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_bar.R
\name{vascr_plot_bar_anova}
\alias{vascr_plot_bar_anova}
\title{Plot a bar chart with ANOVA statistics superimposed on it}
\usage{
vascr_plot_bar_anova(
  data,
  priority,
  confidence,
  time,
  unit,
  frequency,
  format = toplot,
  error = Inf,
  ...
)
}
\arguments{
\item{data}{A vascr dataset}

\item{priority}{The priority of values to export}

\item{confidence}{The minmum confidence level to display}

\item{...}{Any arguement to vascr_prep_graphdata or vascr_polish_plot. Use this to select values ect.}
}
\value{
A vascr bar plot with statistics attached to it
}
\description{
Plot a bar chart with ANOVA statistics superimposed on it
}
\examples{
vascr_plot_bar(data = growth.df, confidence = 0.95, unit = "R", time = 100, frequency = 4000, rotate_x_angle = 45)
vascr_plot_bar_anova(data = growth.df, confidence = 0.95, unit = "R", time = 100, frequency = 4000, rotate_x_angle = 45)

}
