% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_column}
\alias{vascr_plot_column}
\title{Title}
\usage{
vascr_plot_column(
  data,
  replication,
  error = Inf,
  title,
  xlab,
  ylab,
  time,
  unit,
  frequency,
  confidence
)
}
\arguments{
\item{data}{}

\item{replication}{}

\item{error}{}

\item{title}{}

\item{xlab}{}

\item{ylab}{}

\item{time}{}

\item{unit}{}

\item{frequency}{}

\item{confidence}{}
}
\value{

}
\description{
Title
}
