% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_plot_cross_correlation}
\alias{vascr_plot_cross_correlation}
\title{Plot cross correlation data from a vascr dataset}
\usage{
vascr_plot_cross_correlation(
  data.df,
  reference = NULL,
  comparator = NULL,
  manualpairs = NULL
)
}
\arguments{
\item{data.df}{The data set to plot}

\item{reference}{Reference line to compare against, if not specified all comparasons will be made}

\item{comparator}{A list of samples to be compared to the reference. If not specified, all will be compared.}

\item{manualpairs}{A manual list of pairs to be compared}
}
\value{
A plots of the cross correlation calculated
}
\description{
Plot cross correlation data from a vascr dataset
}
\examples{
vascr_plot_cross_correlation(growth.df \%>\% vascr_subset(unit = "R", frequency = 4000))

}
