% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_plot_cross_correlation}
\alias{vascr_plot_cross_correlation}
\title{Plot cross correlation data from a vascr dataset}
\usage{
vascr_plot_cross_correlation(
  data.df,
  reference = NULL,
  comparator = NULL,
  manualpairs = NULL,
  show_index = NULL,
  plot = "all",
  order = TRUE
)
}
\arguments{
\item{data.df}{The dataset to plot}
}
\value{
A plots of the cross correlation calculated
}
\description{
Plot cross correlation data from a vascr dataset
}
\examples{
# vascr_plot_cross_correlation(growth.df)

}
\keyword{internal}
