% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_line}
\alias{vascr_plot_line}
\title{Title}
\usage{
vascr_plot_line(data, priority = NULL)
}
\arguments{
\item{data}{}

\item{replication}{}

\item{error}{}

\item{title}{}

\item{xlab}{}

\item{ylab}{}

\item{linesize}{}

\item{alphavalue}{}
}
\value{

}
\description{
Title
}
\examples{
data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "wells")
vascr_plot_line(data)
vascr_plot_line(data, priority = c("Experiment", "Sample"))

data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "experiments")
data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "summary")


}
