% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_line.R
\name{vascr_plot_line}
\alias{vascr_plot_line}
\title{Title}
\usage{
vascr_plot_line(data.df, errorbars = Inf, alpha = 0.3)
}
\arguments{
\item{data.df}{The vascr data set to plot}

\item{errorbars}{Type of error bars, Inf for ribbons, 0 for no errors and any interger to plot every nth line}

\item{alpha}{Transparency of the error ribbon plotted}
}
\description{
Title
}
