% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_line}
\alias{vascr_plot_line}
\title{Title}
\usage{
vascr_plot_line(data, priority = NULL, error = Inf, alpha = 0.1, ...)
}
\arguments{
\item{data}{}

\item{error}{}

\item{alpha}{}

\item{replication}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "summary", title = "AAA")
vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "experiments", title = "AAA")
vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "wells", title = "AAA")
vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "deviation", title = "AAA")

}
