% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_line.R
\name{vascr_plot_line}
\alias{vascr_plot_line}
\title{Plot a line graph from a vascr dataset}
\usage{
vascr_plot_line(data.df, priority = NULL, error = Inf, alpha = 0.1, ...)
}
\arguments{
\item{data.df}{Vascr dataset to plot}

\item{priority}{Priority list of variables to plot}

\item{error}{Error to plot}

\item{alpha}{Alpha value of the line plotted}

\item{...}{Values to be passed onto prep_graphdata and polish_plot}
}
\value{
A ggplot2 object of the graph created
}
\description{
Plot a line graph from a vascr dataset
}
\examples{
# vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "summary", error = Inf)


# vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "experiments", title = "AAA")
# vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "wells", title = "AAA")
# vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "deviation", title = "AAA")

#vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "summary",
# title = "AAA")
#vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "summary",
# title = "AAA", error = 1)
#vascr_plot_line(growth.df, unit = "R", frequency = 4000, level = "summary",
# title = "AAA", error = 5)

#vascr_plot_line(growth.df, unit = "R", frequency = "raw", time = 50,
# level = "experiments", title = "AAA", error = 0, logscale = "x")

}
\keyword{internal}
