% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_graphics.R
\name{vascr_plot_line}
\alias{vascr_plot_line}
\title{Plot a line graph with vascr}
\usage{
vascr_plot_line(data.df, errorbars = Inf, alpha = 0.3)
}
\arguments{
\item{data.df}{The vascr data set to plot}

\item{errorbars}{Type of error bars, Inf for ribbons, 0 for no errors and any integer to plot every nth line}

\item{alpha}{Transparency of the error ribbon plotted}
}
\value{
a ggplot containing the data
}
\description{
Plot a line graph with vascr
}
\examples{
vascr_plot_line(data.df = growth.df \%>\% vascr_subset(unit = "Rb"))
vascr_plot_line(growth.df \%>\% vascr_subset(unit = "Rb") \%>\% vascr_summarise(level = "experiments"))
data.df = growth.df \%>\% vascr_subset(unit = "Rb") \%>\% vascr_summarise(level = "summary")
vascr_plot_line(data.df)

}
