% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova.R
\name{vascr_plot_line_dunnett}
\alias{vascr_plot_line_dunnett}
\title{Title}
\usage{
vascr_plot_line_dunnett(
  data.df,
  unit = unit,
  frequency = frequency,
  time = time,
  reference = reference,
  inputplot = NULL
)
}
\arguments{
\item{data.df}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}

\item{reference}{}
}
\description{
Title
}
