% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{vascr_plot_plate}
\alias{vascr_plot_plate}
\title{Plot a heatmap of a particular timepoint}
\usage{
vascr_plot_plate(data, priority = NULL, ...)
}
\arguments{
\item{data}{}

\item{time}{}

\item{unit}{}

\item{frequency}{}
}
\value{

}
\description{
Plot a heatmap of a particular timepoint
}
\examples{

vascr_plot_plate(growth.df, time = 100, unit = "Rb")
vascr_plot_plate(growth.df, time = 100, unit = "R", frequency = "4000")
vascr_plot_plate(growth.df, time = 100, unit = "R", frequency = "4000", level = "deviation")
vascr_plot_plate(growth.df, time = 100, unit = "R", frequency = 4000, level = "structure")

}
