% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_plate.R
\name{vascr_plot_plate}
\alias{vascr_plot_plate}
\title{Plot a heatmap of a particular timepoint}
\usage{
vascr_plot_plate(data, priority = NULL, ...)
}
\arguments{
\item{data}{Vascr dataset to plot}

\item{priority}{Vascr priority to plot}

\item{...}{Other values to pass on to data conditioning funcions}
}
\value{
A vascr platemap of the selected data
}
\description{
Plot a heatmap of a particular timepoint
}
\examples{

vascr_plot_plate(growth.df, time = 100, unit = "Rb")
vascr_plot_plate(growth.df, time = 100, unit = "R", frequency = "4000")
vascr_plot_plate(growth.df, time = 100, unit = "R", frequency = "4000", level = "deviation")
vascr_plot_plate(growth.df, time = 100, unit = "R", frequency = 4000, level = "structure")

}
