% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{vascr_plot_resample}
\alias{vascr_plot_resample}
\title{Title}
\usage{
vascr_plot_resample(
  data.df,
  unit = "R",
  frequency = "4000",
  well = "A01",
  newn = 20,
  plot = TRUE
)
}
\arguments{
\item{data.df}{}

\item{unit}{}

\item{frequency}{}

\item{well}{}
}
\description{
Title
}
\examples{
vascr_plot_resample(growth.df)
vascr_plot_resample(growth.df, plot = FALSE)


}
