% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{vascr_plot_resample}
\alias{vascr_plot_resample}
\title{Plot the data resampling process}
\usage{
vascr_plot_resample(
  data.df,
  unit = "R",
  frequency = "4000",
  well = "A01",
  newn = 20,
  plot = TRUE
)
}
\arguments{
\item{data.df}{Dataset to analyse}

\item{unit}{Unit to use, defaults to R}

\item{frequency}{Frequency to use, defaults to 4000}

\item{well}{Well to use, defaults to A01 (or first well in plate)}
}
\value{
A plot showing how well the resampled data conforms to the actual data set
}
\description{
Plot the data resampling process
}
\examples{
vascr_plot_resample(growth.df)
vascr_plot_resample(growth.df, plot = FALSE)


}
