% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_plot_plate.R
\name{vascr_plot_structure}
\alias{vascr_plot_structure}
\title{Plot the plate structure of a datset}
\usage{
vascr_plot_structure(data, title = "Title", stripidentical = TRUE, ...)
}
\arguments{
\item{data}{The datapoint to plot}

\item{title}{Title to be placed on the graph}
}
\value{
A map of the data sampled
}
\description{
Plot the plate structure of a datset
}
\examples{

vascr_plot_structure(growth.df)

}
