% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_prep_graphdata}
\alias{vascr_prep_graphdata}
\title{Prepare a dataset to be graphed by vascar_graph_xxx}
\usage{
vascr_prep_graphdata(
  data,
  unit = "",
  frequency = Inf,
  time = Inf,
  samplecontains = "",
  experiment = "",
  error = Inf,
  alignkey = NULL,
  normtime = NULL,
  divide = FALSE,
  preprocessed = FALSE,
  continuouscontains = NULL,
  stripidentical = TRUE,
  sortkeyincreasing = TRUE,
  level = "",
  errortype = "sem"
)
}
\arguments{
\item{data}{}

\item{unit}{}

\item{frequency}{}

\item{time}{}

\item{samplecontains}{}

\item{experiment}{}

\item{error}{}

\item{alignkey}{}

\item{normtime}{}

\item{divide}{}

\item{preprocessed}{}

\item{continuouscontains}{}

\item{stripidentical}{}
}
\value{

}
\description{
Central data subset, cleanup and label prep function for generation of graphics
}
\examples{

data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "summary")
data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "experiments")
data = vascr_prep_graphdata(growth.df, unit = "Rb", level = "wells")


}
