% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_prep_stat_priority}
\alias{vascr_prep_stat_priority}
\title{Prep priority for statistical data}
\usage{
vascr_prep_stat_priority(data.df, priority = NULL)
}
\arguments{
\item{data.df}{The dataset to use}

\item{priority}{Manual priority (if set)}
}
\value{
A priority list for use in a vascr dataset
}
\description{
Prep priority for statistical data
}
\examples{
#vascr_prep_stat_priority(growth.df)
}
\keyword{internal}
