% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_priority}
\alias{vascr_priority}
\title{Title}
\usage{
vascr_priority(data, explicit = NULL, priority = NULL)
}
\arguments{
\item{data}{}

\item{explicit}{}
}
\value{

}
\description{
Title
}
\examples{


}
