% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{vascr_prism}
\alias{vascr_prism}
\title{Export a prism-compatable data set}
\usage{
vascr_prism(
  data.df,
  unit,
  frequency = 0,
  replication = "summary",
  experiment = ""
)
}
\arguments{
\item{data.df}{ECIS dataframe}

\item{unit}{Unit of data to export}

\item{frequency}{Frequency of data requored, modeled data defaults to 0}

\item{replication}{The level of replicaiton to output. Default is summary of all experiments}

\item{experiment}{The experiment to export. Default is all experiments}
}
\value{
A data frame that can be copied and pasted into prism
}
\description{
This function will automatically average the wells in each experiment, to give a prism-compatable dataset at the experiment level. This allows prism to generate appropriate error bars.
}
\examples{
#vascr_prism(growth.df, 'Rb', replication = "summary")
#prism = vascr_prism(growth.df, 'Rb', replication = "wells")

}
