% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_subset.R
\name{vascr_realise_frequencies}
\alias{vascr_realise_frequencies}
\title{Split out all the frequenices if all is presented}
\usage{
vascr_realise_frequencies(data.df, frequency)
}
\arguments{
\item{data.df}{The dataet to analyse}

\item{frequency}{The frequency to analyse}
}
\value{
A number of a frequency in the dataset
}
\description{
Split out all the frequenices if all is presented
}
\examples{

#vascr_realise_frequencies(growth.df, c(4594, 3000, "all"))

}
\keyword{internal}
