% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_realise_units}
\alias{vascr_realise_units}
\title{Title}
\usage{
vascr_realise_units(data, unit)
}
\arguments{
\item{data}{}

\item{unit}{}
}
\value{

}
\description{
Title
}
\examples{
vascr_titles_vector(c("Rb", "R", "Cm"))

}
