% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_realise_units}
\alias{vascr_realise_units}
\title{Realise vascr units}
\usage{
vascr_realise_units(data, unit)
}
\arguments{
\item{data}{The dataset to process}

\item{unit}{The unit(s) to return}
}
\value{
A vector of vascr units
}
\description{
Realise vascr units
}
\examples{
vascr_realise_units(growth.df, c("Rb", "R", "Cm"))
vascr_realise_units(growth.df, "all")
vascr_realise_units(growth.df, "modeled")
vascr_realise_units(growth.df, "raw")

}
