% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_remove_from_vector}
\alias{vascr_remove_from_vector}
\title{Remove an item from a vector}
\usage{
vascr_remove_from_vector(values, vector)
}
\arguments{
\item{values}{The value(s) to remove}

\item{vector}{The vector to search and modify}
}
\value{
The modified vector
}
\description{
This function cleans up the default R syntax
}
\examples{

vector = unique(growth.df$Unit)
vector = vascr_remove_from_vector("Rb", vector)

}
