% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vascr_rename_columns}
\alias{vascr_rename_columns}
\title{Rename a column - cleanup of the built in funciton}
\usage{
vascr_rename_columns(data.df, old, new)
}
\arguments{
\item{data.df}{The datset to rename}

\item{old}{Old column name}

\item{new}{Replacement column name}
}
\value{
The updated dataset
}
\description{
Rename a column - cleanup of the built in funciton
}
\examples{
# vascr_rename_columns(data.df, "Unit", "Measurement")

}
\keyword{internal}
