% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_align.R
\name{vascr_sample_table}
\alias{vascr_sample_table}
\title{Calculate a table of samples present in a dataset}
\usage{
vascr_sample_table(data.df)
}
\arguments{
\item{data.df}{The dataset to create a table of}
}
\value{
The summarised sample table
}
\description{
Calculate a table of samples present in a dataset
}
\examples{
# data.df = xcell
# samplecols = vascr_sample_table(data.df)
}
\keyword{internal}
