% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_id_work.R
\name{vascr_samples}
\alias{vascr_samples}
\title{Output a list of Sample ID and Sample pairs}
\usage{
vascr_samples(data.df)
}
\arguments{
\item{data.df}{Vascr dataset to handle}
}
\value{
A table of samples and thier corresponding pairs
}
\description{
Output a list of Sample ID and Sample pairs
}
\examples{
# growth.df \%>\% vascr_samples()
}
