% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_save.R
\name{vascr_save}
\alias{vascr_save}
\title{Save a vascr dataset}
\usage{
vascr_save(data.df, path)
}
\arguments{
\item{data.df}{The vascr dataset to save}

\item{path}{The path to save the file to}
}
\value{
A .vascr file containing a vascr dataset
}
\description{
Save a vascr dataset
}
\examples{
path = tempfile()
vascr_save(growth.df, path = path)

}
