% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_shiny.R
\name{vascr_shiny}
\alias{vascr_shiny}
\title{Title}
\usage{
vascr_shiny(data.df = NULL)
}
\arguments{
\item{data.df}{A vascr dataset to maniuplate}
}
\value{
a vascr dataset, with the modifications made
}
\description{
Title
}
\examples{
vascr_shiny(growth.df)
vascr_shiny()
}
