% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_functions.R
\name{vascr_shorten_name}
\alias{vascr_shorten_name}
\title{Title}
\usage{
vascr_shorten_name(
  name_vector,
  select_cols = NULL,
  remove_blank = TRUE,
  fill_blank = "Control",
  include_wells = FALSE
)
}
\arguments{
\item{name_vector}{}

\item{select_cols}{}

\item{remove_blank}{}

\item{fill_blank}{}
}
\value{

}
\description{
Title
}
