% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{vascr_subset_continuous}
\alias{vascr_subset_continuous}
\title{Subset a continuous variable}
\usage{
vascr_subset_continuous(
  data,
  continuous,
  exact_match = FALSE,
  strip_empty = TRUE,
  implode = TRUE
)
}
\arguments{
\item{data}{A vascr dataset to subset}

\item{continuous}{The continuous variable to subset}

\item{exact_match}{Should the variables selected have an exact match to the column names input, or only contain the value input. Default is containing as otherwise you have to keep track of units.}

\item{strip_empty}{Should columns in which none of the selected variables are present be removed}

\item{implode}{Should the final data set be imploded, replacing the sample wells present}
}
\value{
A vascr dataset subsampled on a continuous variable
}
\description{
Subset the columns that are exploded out of a continuous variable. Contains options to remove descriptors that are now defunct, so this will be repaired later.
}
\examples{
# Sub code for breaking out continuous datasets
#exploded = vascr_explode(xcell)
#subset = vascr_subset_continuous(exploded, continuous = "ATP", strip_empty = FALSE)
#vascr_plot(exploded, unit = "CI", frequency = "10000", replication = "experiments"
# , normtime = 160, continuouscontains = "ATP")


}
\keyword{internal}
