% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_id_work.R
\name{vascr_subset_sampleid}
\alias{vascr_subset_sampleid}
\title{#' Sub setting function for sample IDs
#'
#' @param data.df 
#' @param samplelist 
#'
#' @return A subset data frame
#' @noRd
#'
#' @examples}
\usage{
vascr_subset_sampleid(data.df, samplelist)
}
\description{
vascr_subset_sampleid = function(data.df, samplelist)
{
  
  ided = data.df %>% select(Sample) %>% distinct() %>% mutate(SampleID = row_number()) %>% 
    right_join(data.df, by = "Sample")
  
  toreturn = ided %>% subset(SampleID %in% samplelist)
  
  return(toreturn)
  
}
}
