% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_summarise_change}
\alias{vascr_summarise_change}
\title{Create a data frame summarising the changing variables in a dataset in a single column}
\usage{
vascr_summarise_change(data.df)
}
\arguments{
\item{data.df}{The dataset to summarise}
}
\value{
A full dataset
}
\description{
Create a data frame summarising the changing variables in a dataset in a single column
}
\examples{
data.df = growth.df
dat = vascr_summarise_change(data.df)

}
\keyword{internal}
