% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{vascr_summarise_cross_correlation}
\alias{vascr_summarise_cross_correlation}
\title{Title}
\usage{
vascr_summarise_cross_correlation(data.df)
}
\arguments{
\item{data.df}{}
}
\value{

}
\description{
Title
}
\examples{
growthrb = vascr_subset(growth.df, unit = "Rb", time = c(5,150))
vascr_plot(growthrb, unit ="Rb")
vascr_summarise_cross_correlation(growthrb)

}
