% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tests.R
\name{vascr_test_normalised}
\alias{vascr_test_normalised}
\title{Test if a dataframe is normalised}
\usage{
vascr_test_normalised(data.df)
}
\arguments{
\item{data.df}{A vascr dataframe}
}
\value{
TRUE if normalised, FALSE if not
}
\description{
Test if a dataframe is normalised
}
\examples{
vascr_test_normalised(growth.df)

normdata  = vascr_normalise(growth.df, 100)

vascr_test_normalised(normdata)

}
