% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tests.R
\name{vascr_test_summary_level}
\alias{vascr_test_summary_level}
\title{Test the summary level of an ECIS dataframe}
\usage{
vascr_test_summary_level(data.df)
}
\arguments{
\item{data.df}{A standard ECIS dataframe}
}
\value{
A summarised standard ECIS dataframe
}
\description{
Test the summary level of an ECIS dataframe
}
\examples{
vascr_test_summary_level(growth.df)

}
