% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascr_anova_new.R
\name{vascr_tukey}
\alias{vascr_tukey}
\title{Generate a tukey analyasis of vascr data}
\usage{
vascr_tukey(data.df, unit, frequency, time, priority = NULL, raw = FALSE)
}
\arguments{
\item{data.df}{The dataset to analyse}

\item{unit}{Unit to analyse}

\item{frequency}{Frequency to analyse}

\item{time}{Time to analyse}

\item{priority}{Priority of units for statistical analyasis}

\item{raw}{If true, a non-processed form of the tukey results will be returned}
}
\value{
A table or Tukey HSD test result
}
\description{
Generate a tukey analyasis of vascr data
}
\examples{
# vascr_tukey(growth.df, "R", 4000, 100)
# vascr_tukey(growth.df, "R", 4000, 100, raw = TRUE)

}
\keyword{internal}
