% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{vascr_zero_time}
\alias{vascr_zero_time}
\title{Set a particular time point as 0 in a vascr dataset}
\usage{
vascr_zero_time(data.df, time = 0)
}
\arguments{
\item{data.df}{A vascr dataset}

\item{time}{The time to be set to 0}
}
\value{
An adjusted vascr dataset
}
\description{
Allows the user to change the time designated as zero to allow clearer plotting of treatments
}
\examples{
vascr_zero_time(growth.df, 50)

}
